% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Batch_Effect_Correction}
\alias{Batch_Effect_Correction}
\title{Batch effect correction using ComBat from the sva package.}
\usage{
Batch_Effect_Correction(OmicSampleData, Metadata, BatchColumn)
}
\arguments{
\item{OmicSampleData}{Single omics dataset, with features in rows.}

\item{Metadata}{single omics metadata file, with batch information.}

\item{BatchColumn}{String of the column name which contains batch information
in the metadata file.}
}
\value{
Single omics matrix with sample batch effect corrected.
}
\description{
Batch effect correction using ComBat from the sva package. Only possible if
batch information is available for the samples. Used in EINS R6 class
function \code{run_Preprocessing}, if \code{FunctionOrder} includes \code{"BatchCorrection"}.
}
\references{
Leek JT, Johnson WE, Parker HS, Fertig EJ, Jaffe AE, Zhang Y,
Storey JD, Torres LC (2025). sva: Surrogate Variable Analysis.
doi:10.18129/B9.bioc.sva, R package version 3.56.0
}
\keyword{internal}
