% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Concat_Feature_Matrix}
\alias{Concat_Feature_Matrix}
\title{Concatenate feature weight results of all multi-omics integration methods
into a single matrix.}
\usage{
Concat_Feature_Matrix(ScaledFeaturesData, nFactors = NULL)
}
\arguments{
\item{ScaledFeaturesData}{List of scaled feature weight dataframes for the
different multi-omics integration methods, by number of factors/components
and omics type, as calculated with \code{Feature_Weight_Scaling()}.}

\item{nFactors}{Number of factors (components) for which the hierarchical
clustering is to be performed. All multi-omics integration methods will need
to have been performed with this number of factors/components.}
}
\value{
A single matrix of scaled feature weights. Features in rows, factors
or components per multi-omics integration method in columns.
}
\description{
Create a single matrix by concatenating the scaled feature weight results as
calculated in \code{Feature_Weight_Scaling()}, with features in rows and factors
(components) per multi-omics integration method in columns. To be used in
\code{Feature_Concat_HClust()}.
}
\keyword{internal}
