% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{CorFilter}
\alias{CorFilter}
\title{Filter out correlated features}
\usage{
CorFilter(OmicSampleData, CutOff = 0.9, CorMean = FALSE)
}
\arguments{
\item{OmicSampleData}{Single omics dataset, with features in rows.}

\item{CutOff}{Numerical threshold for correlation. Default is 0.9 (if
pairwise correlation between 2 features is 0.9 or larger, they are filtered).}

\item{CorMean}{Whether to calculate the mean measurements for the correlated
features. Default is \code{FALSE}. If \code{TRUE}, the mean of all feature measurements
for a group of correlated features is included in the dataset.}
}
\value{
Single omics matrix, with correlated features filtered out
}
\description{
Filter out correlated features from the omics dataset. Calculates pairwise
correlations between features, and filter out those features with correlation
above a user-defined threshold. Either one feature is kept as the
representative, with the filtered out features in a list by representative,
or the mean measurements from the correlated features is calculated and kept.
Used in EINS R6 class function \code{run_Preprocessing}, if \code{FunctionOrder}
includes \code{"FilterCorrelation"}.
}
\keyword{internal}
