% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Data_Manipulation_Sankey_Methods}
\alias{Data_Manipulation_Sankey_Methods}
\title{Create dataframe to be used for Sankey plot comparing multi-omics integration
methods}
\usage{
Data_Manipulation_Sankey_Methods(ClusterResData, MetadataColumn)
}
\arguments{
\item{ClusterResData}{List of cluster assignment dataframes.}

\item{MetadataColumn}{String of the sample metadata feature for which the
samples should be colored. The metadata feature needs to be available
in the metadata files for all omics types.}
}
\value{
List with results. $Links, dataframe with samples in rows, and
columns indicating source cluster and target cluster, as well as metadata
color. $Nodes, dataframe with source and target clusters.
}
\description{
Combine the cluster results for multiple multi-omics methods performed for
a single cluster number in a single dataframe. For each sample with each
different cluster assignment, the sample is given a source and target value,
which can be used to display the flow in the Sankey plot. Each sample also
gets a color assigned based on a user-defined metadata feature.
}
\keyword{internal}
