% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Dendrogram_Label_Groups}
\alias{Dendrogram_Label_Groups}
\title{Create a list of dataframes with label coloring and legend information for a
dendrogram}
\usage{
Dendrogram_Label_Groups(Feature, HClust, Dendrogram, LabelSize)
}
\arguments{
\item{Feature}{Whether feature or factor dendrogram is being used. If \code{TRUE},
the input dendrogram is a feature dendrogram. If \code{FALSE}, the input
dendrogram is a factor dendrogram.}

\item{HClust}{Hierarchical clustering trees.}

\item{Dendrogram}{Dendrogram.}

\item{LabelSize}{Whether to alternate label size in feature dendrogram, to
distinguish the omics.}
}
\value{
List of dataframes. $Colors, dataframe with dendrogram labels and
the colors to be used. $Legend_Levels, dataframe with dendrogram label levels.
$Legend_Colors, dataframe with dendrogram label level colors.
}
\description{
Create a list of dataframes with dendrogram labels colored and legend
information for the label colors. To be used with a dendrogram.
}
\keyword{internal}
