% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Dendrogram_Omics_Bar}
\alias{Dendrogram_Omics_Bar}
\title{Create list of dataframes with data for colored bars and legend for
dendrogram}
\usage{
Dendrogram_Omics_Bar(HClust, Dendrogram)
}
\arguments{
\item{HClust}{Hierachical clustering tree.}

\item{Dendrogram}{Dendrogram.}
}
\value{
A list. $Colors, dataframe with the color to be used per feature
label. $Legend_Levels, dataframe with dendrogram omics levels. $Legend_Colors,
dataframe with dendrogram omics level colors. $Vector, vector with color to
be used per feature, ordered by dendrogram order.
}
\description{
Create a list of dataframes with omics information for the colored bars and
legend data for feature dendrograms.
}
\keyword{internal}
