% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Dendrogram_Sample_Meta}
\alias{Dendrogram_Sample_Meta}
\title{Create list of dataframes with sample metadata information and color for
colored bars below dendrograms}
\usage{
Dendrogram_Sample_Meta(MetadataFeatures, Metadata, Dendrogram, LabelMetadata)
}
\arguments{
\item{MetadataFeatures}{Names of the sample metadata features to be
included in colored bars. A string vector with the names of the metadata
features as in the metadata files should be provided.}

\item{Metadata}{Dataframe with sample metadata.}

\item{Dendrogram}{Sample dendrogram.}

\item{LabelMetadata}{Name of the metadata feature for which the sample
labels should be colored. Must be a sting with the name of the metadata
feature as in the metadata files.}
}
\value{
List of dataframes. $Vector, dataframe with colored levels of
metadata features for the colored bars. $Label, dataframe with labels colored
by metadata features, ordered by the dendrogram order.
}
\description{
Create list of dataframes with sample metadata information and colors. These
can be used for create colored bars and colored labels on sample dendrograms.
}
\keyword{internal}
