% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS-package.R
\docType{package}
\name{EINS-package}
\alias{EINS}
\alias{EINS-package}
\title{EINS: Ensemble of Multiple Multi-Omics Integration Methods}
\description{
A package to create an R6 object in which multi-omics integration can be performed using nine methods, and the results combined into ensemble integration findings. Multiple omics datasets can be uploaded in the R6 object, after which single-omics analysis, multi-omics integration, ensemble integration and result visualization can be performed.
}
\author{
\strong{Maintainer}: Mara Zuiderwijk \email{m.o.zuiderwijk@lumc.nl}

Other contributors:
\itemize{
  \item Damien Olivier [contributor]
  \item Yassene Mohammed [contributor]
}

}
\keyword{internal}
