% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Ensemble_Cluster_CHC}
\alias{Ensemble_Cluster_CHC}
\title{Create an ensemble clustering assignment from all available multi-omics
integration results}
\usage{
Ensemble_Cluster_CHC(
  MethodClusterResults,
  Sample_IDs,
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2",
  Clusters = NULL
)
}
\arguments{
\item{MethodClusterResults}{A list of dataframes with cluster assignments per
multi-omics integration method.}

\item{Sample_IDs}{Vector of sample names.}

\item{Distance}{Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{MinkowskiPower}{Power of the Minkowski distance. Default is \code{NULL}.}

\item{Linkage}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"} (default), \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}

\item{Clusters}{Number of clusters to create from the hierarchical clustering
result. Default is \code{NULL}.}
}
\value{
A list of results. Cluster assignment from the hierarchical
clustering result, into the number of cluster as provided in \code{Clusters},
stored in $ClusterRes. Sample distance matrix, stored in $DistMat.
hierarchical clustering tree, stored in $HClustRes.
}
\description{
Using the dataframe with multi-omics integration cluster assignments as
created in \code{Ensemble_Cluster_Data()}, create a sample similarity matrix by
counting the number of times samples cluster together. Hierarchical
clustering is then performed on this matrix to provide a more robust
clustering result by combining multiple cluster assignments.
}
\keyword{internal}
