% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Ensemble_Cluster_Data}
\alias{Ensemble_Cluster_Data}
\title{Prepare the cluster results from the different multi-omics integration
methods for ensemble clustering}
\usage{
Ensemble_Cluster_Data(Data)
}
\arguments{
\item{Data}{A list of dataframes, stored by multi-omics integration method
and cluster number. Each dataframe contains a single cluster result.}
}
\value{
A list of dataframes with samples in rows and cluster assignment
in columns per multi-omics integration method.
}
\description{
Create a list of dataframes with the sample clustering results of the
multi-omics integration methods with all calculated cluster numbers. This
dataframe is used for ensemble clustering in \code{Ensemble_Cluster()}.
}
\keyword{internal}
