% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Ensemble_Cluster_MDS_HC}
\alias{Ensemble_Cluster_MDS_HC}
\title{Create an ensemble clustering assignment from all available multi-omics
integration results and perform Multidimensional Scaling (MDS) followed
by hierarchical clustering}
\usage{
Ensemble_Cluster_MDS_HC(
  MethodClusterResults,
  Sample_IDs,
  Distance = "euclidean",
  MinkowskiPower = NULL,
  CandidateDimensions = c(2:10),
  StabilityEpsilon = 0.02,
  AddConstant = TRUE,
  Linkage = "ward.D2",
  Clusters = NULL
)
}
\arguments{
\item{MethodClusterResults}{A list of dataframes with cluster assignments per
multi-omics integration method.}

\item{Sample_IDs}{Vector of sample names.}

\item{Distance}{Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{MinkowskiPower}{Power of the Minkowski distance. Default is \code{NULL}.}

\item{CandidateDimensions}{Number of MDS dimensions to be tested. Default is
2:10 dimensions.}

\item{StabilityEpsilon}{How much the selected leave-one-method-out run may
diverge from the maximum observed stability, default is 0.02.}

\item{AddConstant}{Logical, indicating if an additive constant c* should
be computed, and added to the non-diagonal dissimilarities such that the
modified dissimilarities are Euclidean.}

\item{Linkage}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"} (default), \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}

\item{Clusters}{Number of clusters to create from the hierarchical clustering
result. Default is \code{NULL}.}
}
\value{
A list of results. Sample distance matrix, stored in $DistMat.
Consensus matrix, stored in $ConsensusMatrix. Dissimilarity matrix, stored in
$DissimilarityMatrix. MDS final embedding, stored in $Embed. MDS embedding
distance matrix, $EmbedDist. Full final MDS result, provided in $MDS.
Dimension calculation information, provided in $DimensionInfo. Hierarchical
clustering tree, stored in $HClustRes.
}
\description{
Using the dataframe with multi-omics integration cluster assignments as
created in \code{Ensemble_Cluster_Data()}, create a sample similarity matrix by
counting the number of times samples cluster together. Classical MDS is
applied to this matrix to obtain sample coordinates. Hierarchical
clustering is then performed on this MDS matrix. The MDS dimension is
selected automatically, by using leave-one-method-out stabilities, to favor
the most compact representation which preserves robustness. Stability is
quantified as the Adjusted Rand Index  between the full-ensemble MDS-HC
clustering and each leave-one-method-out MDS-HC clustering.
EnsembleClustering = list()
}
\keyword{internal}
