% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_R6_v5.R
\name{Ensemble_Integration_Suite}
\alias{Ensemble_Integration_Suite}
\title{Ensemble_Integration_Suite Object.}
\value{
Raw data stored in the object under $Omics$Raw_Data, metadata
stored in the object under $Omics$Metadata

All $Omics and $Preprocessed_Omics files for the omics defined
in \code{OmicsName} will become \code{NULL}, removing the matrices.

Raw data stored in object under $Omics$Raw_Data, and metadata
stored in object under $Omics$Metadata.

$Omics$Raw_Data matrices with samples stored as columns.

Preprocessed data matrix for the defined omics type. Stored in
$Preprocessed_Omics$Data

If the samples do not match across all omics datasets, only the
matching samples are stored in $Preprocessed_Omics$Matched_Data. The
matrix columns are ordered the same way across omics datasets and
metadata.

A list of the optimal cluster number calculated by each included
index for each integration method per component/factor. The mean, median
and mode are calculated per integration method, and for all methods
combined.

The arguments and fit of the model for each omics dataset,
stored in $Single_Omics$Fit. A dataframe with the cluster assignment for
each omics dataset, stored in $Single_Omics$ClusterRes.

Distance matrix for each individual omics dataset, stored in
$Single_Omics$DistanceMatrix. Hclust object for each individual omics
dataset, stored in $Single_Omics$HClustTree. Dendrogram for each
individual omics dataset, stored in $Plots$Dendrogram$Single_Omics.

A list of results. Fit of the integration for each included
cluster number, stored in $Multi_Omics$Fit$MoCluster. Sample cluster
assignment for each included cluster number, stored in
$Multi_Omics$ClusterRes$MoCluster. Loading vectors (feature weights)
calculated for each omics dataset, stored by component number in
$Multi_Omics$FeatureRes$MoCluster. Variance in the sample clusters
explained by each component for each included cluster number, stored in
$Multi_Omics$VarianceExp$MoCluster. Sample factor score for each
included cluster number, stored in $Multi_Omics$CoordData$MoCluster.

A list of results. Fit of the integration, stored by component
number in $Multi_Omics$Fit$MCIA. If PerformCluster = \code{TRUE}, sample cluster
assignment for each included cluster number, stored in
$Multi_Omics$ClusterRes$MCIA Loading vectors (feature weights)
calculated for each omics dataset, stored by component number in
$Multi_Omics$FeatureRes$MCIA. Sample factor score, stored by component
number in $Multi_Omics$CoordData$MCIA.

A list of results. Fit of the integration for each included
cluster number, stored in $Multi_Omics$Fit$jNMF. Sample cluster
assignments for each included cluster number, stored in
$Multi_Omics$ClusterRes$jNMF. Component (factor) by feature matrices for
each individual omics dataset, stored by included cluster number
(= components in jNMF) in $Multi_Omics$FeatureRes$jNMF.

A list of results. Fit of the integration for each included
cluster number, stored in $Multi_Omics$Fit$iNMF. Sample cluster
assignments for each included cluster number, stored in
$Multi_Omics$ClusterRes$iNMF. Component (factor) by feature matrices for
each individual omics dataset, stroled by included cluster number
(= components in iNMF) in $Multi_Omics$FeatureRes$iNMF.

A list of results. Fit of the integration for each included
cluster number, stored in $Multi_Omics$Fit$LRAcluster. Sample cluster
assignments for each included cluster number, stored in
$Multi_Omics$ClusterRes$LRAcluster. Matrix with coordinates of all
samples in reduced space for each included cluster number, stored in
$Multi_Omics$CoordData$LRAcluster.

A list of results. Fit of the integration for each included
\code{FullClusters} cluster number, stored in $Multi_Omics$Fit$COCA. Sample
cluster assignments for each included \code{FullClusters} cluster number,
stored in $Multi_Omics$ClusterRes$COCA.

A list of results. Fit of the integration model, both pretrained
and trained, stored by number of factors in $Multi_Omics$Fit$MOFA. Factor
(component) by feature matrices with feature weights for each omics
dataset, stored by number of factors in $Multi_Omics$FeatureRes$MOFA.
Sample by factor (component) matrix, stored by number of factors in
$Multi_Omics$FactorRes$MOFA. Total variance explained for each factor
(component) per omics type, stored by number of factors in
$Multi_Omics$VarianceExp$MOFA. Matrix with coordinates of all
samples in reduced factor (component) space, stored by number of factors
in $Multi_Omics$CoordData$MOFA.

A list of results. Fit of the integration for each included
cluster number and distance metric, stored in $Multi_Omics$Fit$SNF.
Sample cluster assignments for each included cluster number and distance
metric, stored in $Multi_Omics$ClusterRes$SNF.

A list of results. Fit of the single omics UMAPs, concatenated
UMAP and clustering method for each cluster number, stored by clustering
method in $Multi_Omics$Fit$GAUDI. Sample cluster assignments for each
cluster number, stored by clustering method in
$Multi_Omics$ClusterRes$GAUDI. Sample coordinates in the concatenated
UMAP embedding for each cluster number, stored by clustering method in
$Multi_Omics$CoordData$GAUDI.

A list of results. Feature distance matrix, stored by
multi-omics integration method in
$Multi_Omics$Feature_HClust$DistanceMatrix. Feature hierarchical
clustering trees, stored by multi-omics integration method in
$Multi_Omics$Feature_HClust$HClust. Feature dendrogram, stored by
multi-omics integration method in $Plots$Dendrogram$Features$Multi_Omics.

A list of results. Sample distance matrix is stored in
$Ensemble$Samples$MDS$DistanceMatrix by cluster number. Sample hierarchical
clustering tree is stored in $Ensemble$Samples$MDS$HClust by cluster number.
Ensemble sample cluster assignment is stored in
$Ensemble$Samples$MDS$ClusterRes by cluster number. Sample dendrogram is
stored in $Plots$Dendrogram$Ensemble$MDS by cluster number.

A list of results. CCA aligned embedding is stored in
$Ensemble$Samples$CCA$Embedding by factor number. Ensemble sample cluster
assignment is stored in $Ensemble$Samples$CCA$ClusterRes by factor and
cluster number.

A list of results. Feature distance matrix, stored by feature
ensemble method in $Ensemble$Features[\code{\link{Method}}]$DistanceMatrix.
Feature hierachical clustering tree, stored by feature ensemble method in
$Ensemble$Features[\code{\link{Method}}]$HClust. Feature dendrogram, stored by
feature ensemble method in $Plots$Dendrogram$Features. Factor distance
matrix, stored in $Ensemble$Factors$Concatenation$DistanceMatrix. Only
available for \code{Method} \code{"Concatenation"}. Factor hierarchical clustering
tree, stored in $Ensemble$Factors$Concatenation$HClust. Only available
for \code{Method} \code{"Concatenation"}. Factor dendrogram, stored in
$Plots$Dendrogram$Factors$Concatenation. Only available for \code{Method}
\code{"Concatenation"}.

The optimal number of feature clusters, stored in
$Ensemble$Features[\code{\link{Method}}]$OptimalClusters by
\code{run_Feature_Ensemble_HClust} \code{Method} as available.

The optimal number of factor clusters, stored in
$Ensemble$Factors$Concatenation$OptimalClusters.

A list of results. Feature clustering result, stored by
\code{run_Feature_Ensemble_HClust} \code{Method} in
$Ensemble$Features[\code{\link{Method}}]$ClusterRes. Feature dendrogram with
cluster assignment, stored by \code{run_Feature_Ensemble_HClust} \code{Method} in
$Plots$Dendrogram$Features.

A list of results. Factor clustering result, stored in
$Ensemble$Features$Concatenation$ClusterRes. Factor dendrogram with
cluster assignment, stored in $Plots$Dendrogram$Factors$Concatenation.

A list of results. A dataframe with all features of the omics
type per cluster of ensembled features. Stored by \code{Method} in
$Ensemble$Features[\code{\link{Method}}]$ORA[\code{\link{OmicsName}}][\code{\link{Clusters}}]$FeatureList.
Enrichment result of class \code{enrichResult} per cluster of ensembled
features, if that cluster contains features of the selected omics type
and enrichment analysis resulted in enriched terms. Stored by \code{Method} in
$Ensemble$Features[\code{\link{Method}}]$ORA[\code{\link{OmicsName}}][\code{\link{Clusters}}]$Enrichment.

A list of plots. Heatmap with cluster assignment for each omics
type per multi-omics or ensemble integration clustering results. Stored
by integration method and cluster number for each omics type in
$Plots$Cluster_Heatmap.

A list of plots. Multiple omics heatmaps with cluster assignment
and metadata features per multi-omics or ensemble integration clustering
result. Stored by integration method and cluster number in
$Plots$Multi_Omics_Heatmap

A list of plots. Interactive scatterplots, with dots colored by
cluster assignment and shaped by \code{MetadataFeature}. Stored by integration
method and cluster number in $Plots$Scatterplot

A list of plots. Cluster assignment and metadata feature table,
stored by integration method and cluster number in
$Plots$Clusters_Metadata.

A list of plots. Sankey plot with the results of multiple
cluster assignments from a single multi-omics integration method. Stored
by multi-omics integration method in $Plots$Sankey_Clusters.

A list of plots. Sankey plot with the cluster assignments of
all multi-omics integration methods and the specified \code{SNFDistance},
stored by cluster number in $Plots$Sankey_Methods$Selected. If \code{AllPlots}
was \code{TRUE}, Sankey plot with the cluster assignments of all calculated
SNF distances, stored by cluster number in $Plots$Sankey_Methods$SNF. If
\code{AllPlots} was \code{TRUE}, Sankey plot with the cluster assignment of all
multi-omics integration methods and all SNF distances, stored by cluster
number in $Plots$Sankey_Methods$Full.

A list of plots. Top features for each factor (component),
stored by multi-omics integration method, number of factors and omics
type in Plots$Top_Feature_Weights.

A single or multiple dendrograms. If \code{Data} = \code{"Single"}, a
sample dendrogram per omics type is returned, with colored bars with
metadata features below each dendrogram. If \code{Data} = \code{"Ensemble MDS"},
a sample dendrogram is returned, with colored bars with metadata features
below the dendrogram. If \code{Data} = \code{"Ensemble Factor"}, a factor
dendrogram is returned. If \code{Data} =  \code{"Ensemble Feature Average"}, a
feature dendrogram is returned, with a colored bar indicating the feature
omics type below the dendrogram. If \code{Data} =
\code{"Ensemble Feature Concatenation"}, a feature dendrogram is returned,
with a colored bar indicating the feature omics type below the dendrogram.
If \code{Data} = \code{"Multi-Omics Feature"}, a feature dendrogram per multi-omics
integration method is returned, with colored bars indicating the feature
omics type below each dendrogram.

A list of plots. A dotplot with the overrepresented GO terms per
omics type for each cluster, stored by \code{Method}, omics type and cluster
number in $Plots$ORA_Dot.
}
\description{
An object in which to import multiple omics datasets, which can then be
analyzed with single omics, multi-omics integration and ensemble integration
methods. Visualization methods are included.
}
\details{
At least 2 different omics datasets are required, with metadata files. Single
omics analysis is possible with model-based clustering or hierarchical
clustering. 9 different multi-omics integration methods are available:
MoCluster, MCIA, jNMF, iNMF, LRAcluster, MOFA, COCA, GAUDI and SNF. After
multi-omics analysis, ensemble sample clustering and ensemble feature
clustering can be performed using the multi-omics integration results.
Visualization of the clustering results can then be performed.
}
\examples{

## ------------------------------------------------
## Method `Ensemble_Integration_Suite$run_Sample_Matching`
## ------------------------------------------------

E <- Ensemble_Integration_Suite$new()
E$add_Omics_df(OmicsName = "Proteomics",
DataDF = EINS_NCI60_Toy$Proteomics, MetaDF = EINS_NCI60_Toy$Metadata)
E$add_Omics_df(OmicsName = "Methylation",
DataDF = EINS_NCI60_Toy$Methylation,
MetaDF = EINS_NCI60_Toy$Methylation)=
E$run_Sample_Matching()
E$Preprocessed_Omics$Matched_Data$Proteomics

}
\references{
\itemize{
\item Hastie T, Tibshirani R, Narasimhan B, Chu G (2025). impute:
impute: Imputation for microarray data. doi:10.18129/B9.bioc.impute,
R package version 1.82.0
\itemize{
\item Stekhoven DJ (2022). missForest: Nonparametric Missing Value Imputation
using Random Forest. R package version 1.5.
\item Leek JT, Johnson WE, Parker HS, Fertig EJ, Jaffe AE, Zhang Y, Storey
JD, Torres LC (2025). sva: Surrogate Variable Analysis.
doi:10.18129/B9.bioc.sva, R package version 3.56.0
}
}

\itemize{
\item Charrad M, Ghazzali N, Boiteau V, Niknafs A (2014).
“NbClust: An R Package for Determining the Relevant Number of Clusters
in a Data Set.” Journal of Statistical Software, 61(6), 1–36.
\item Meng C (2025). mogsa: Multiple omics data integrative clustering and
gene set analysis. doi:10.18129/B9.bioc.mogsa, R package version 1.42.0
\item Meng C, Kuster B, Culhane A, Gholami AM (2013). “A multivariate
approach to the integration of multi-omics datasets.” BMC Bioinformatics.
\item Tsuyuzaki, K., and Nikaido, I. (2024). nnTensor: Non-Negative Tensor
Decomposition. R package version 1.3.0
\item Chalise, P., Raghavan, R., and Fridley, B. (2025). IntNMF: Integrative
Clustering of Multiple Genomic Dataset. R package version 1.3.0
\item Lu, X., Meng, J., Zhou, Y., Jiang, L., and Yan, F. (2020).MOVICS: an R
package for multi-omics integration and visualization in cancer
subtyping. Bioinformatics, btaa1018.
\item Wu D, Wang D, Zhang MQ, Gu J (2015). Fast dimension reduction and
integrative clustering of multi-omics data using low-rank approximation:
application to cancer molecular classification. BMC Genomics, 16(1):1022.
\item Argelaguet R, Velten B, Arnol D, Dietrich S, Zenz T, Marioni JC,
Buettner F, Huber W, Stegle O (2018). “Multi‐Omics Factor Analysis—a
framework for unsupervised integration of multi‐omics data sets.”
Molecular Systems Biology, 14.
\item Castellano-Escuder P, Zachman DK, Han K, Hirschey MD. GAUDI:
interpretable multi-omics integration with UMAP embeddings and
density-based clustering. Nat Commun. 2025 Jul 1;16(1):5771.
}

Bergé L, Bouveyron C, Girard S (2012). “HDclassif: An R
Package for Model-Based Clustering and Discriminant Analysis of
High-Dimensional Data.” Journal of Statistical Software, 46(6), 1–29.

Meng C (2025). mogsa: Multiple omics data integrative
clustering and gene set analysis. doi:10.18129/B9.bioc.mogsa, R package
version 1.42.0

Meng C, Kuster B, Culhane A, Gholami AM (2013). “A
multivariate approach to the integration of multi-omics datasets.” BMC
Bioinformatics.

Tsuyuzaki, K., and Nikaido, I. (2024). nnTensor: Non-Negative
Tensor Decomposition. R package version 1.3.0

Chalise, P., Raghavan, R., and Fridley, B. (2025). IntNMF:
Integrative Clustering of Multiple Genomic Dataset. R package version
1.3.0

Wu D, Wang D, Zhang MQ, Gu J (2015). Fast dimension reduction
and integrative clustering of multi-omics data using low-rank approximation:
application to cancer molecular classification. BMC Genomics, 16(1):1022.

\itemize{
\item Lu, X., Meng, J., Zhou, Y., Jiang, L., and Yan, F. (2020).
MOVICS: an R package for multi-omics integration and visualization in
cancer subtyping. Bioinformatics, btaa1018.
\item Wu D, Wang D, Zhang MQ, Gu J (2015). Fast dimension reduction and
integrative clustering of multi-omics data using low-rank approximation:
application to cancer molecular classification. BMC Genomics, 16(1):1022.
}

Cabassi A, Kirk PD (2020). “Multiple kernel learning for
integrative consensus clustering of genomic datasets.” Bioinformatics.

Argelaguet R, Velten B, Arnol D, Dietrich S, Zenz T, Marioni
JC, Buettner F, Huber W, Stegle O (2018). “Multi‐Omics Factor Analysis—a
framework for unsupervised integration of multi‐omics data sets.”
Molecular Systems Biology, 14.

Wang, B., Mezlini, A., Demir, F., Fiume, M., Tu, Z., Brudno,
M., Haibe-Kains, B., and Goldenberg, A. (2021) SNFtool: Similarity
Network Fusion. R package version 2.3.1

\itemize{
\item Castellano-Escuder P, Zachman DK, Han K, Hirschey MD.
GAUDI: interpretable multi-omics integration with UMAP embeddings and
density-based clustering. Nat Commun. 2025 Jul 1;16(1):5771.
\item Hahsler M, Piekenbrock M (2025). dbscan: Density-Based Spatial
Clustering of Applications with Noise (DBSCAN) and Related Algorithms.
doi:10.32614/CRAN.package.dbscan, R package version 1.2.3
}

\itemize{
\item Meng C (2025). mogsa: Multiple omics data integrative
clustering and gene set analysis. doi:10.18129/B9.bioc.mogsa, R package
version 1.42.0
\item Meng C, Kuster B, Culhane A, Gholami AM (2013). “A multivariate
approach to the integration of multi-omics datasets.” BMC Bioinformatics.
\item Tsuyuzaki, K., and Nikaido, I. (2024). nnTensor: Non-Negative Tensor
Decomposition. R package version 1.3.0
\item Chalise, P., Raghavan, R., and Fridley, B. (2025). IntNMF: Integrative
Clustering of Multiple Genomic Dataset. R package version 1.3.0
\item Argelaguet R, Velten B, Arnol D, Dietrich S, Zenz T, Marioni JC,
Buettner F, Huber W, Stegle O (2018). “Multi‐Omics Factor Analysis—a
framework for unsupervised integration of multi‐omics data sets.”
Molecular Systems Biology, 14.
}

Studer M (2013). “WeightedCluster Library Manual: A practical
guide to creating typologies of trajectories in the social sciences with
R.” LIVES Working Papers 24.

Studer M (2013). “WeightedCluster Library Manual: A practical
guide to creating typologies of trajectories in the social sciences with
R.” LIVES Working Papers 24.

\itemize{
\item Yu G (2024). “Thirteen years of clusterProfiler.” The
Innovation, 5(6), 100722.
\item Xu S, Hu E, Cai Y, Xie Z, Luo X, Zhan L, Tang W, Wang Q, Liu B, Wang R,
Xie W, Wu T, Xie L, Yu G (2024). “Using clusterProfiler to characterize
multiomics data.” Nature Protocols, 19(11).
\item Wu T, Hu E, Xu S, Chen M, Guo P, Dai Z, Feng T, Zhou L, Tang W, Zhan L,
Fu x, Liu S, Bo X, Yu G (2021). “clusterProfiler 4.0: A universal
enrichment tool for interpreting omics data.” The Innovation, 2(3).
\item Yu G, Wang L, Han Y, He Q (2012). “clusterProfiler: an R package for
comparing biological themes among gene clusters.” OMICS: A Journal of
Integrative Biology, 16(5), 284-287.
\item Carlson M (2025). "Genome wide annotation for Human, primarily based on
mapping using Entrez Gene identifiers." Bioconductor.
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Omics}}{List of different datasets per omics type.}

\item{\code{Preprocessed_Omics}}{List of datasets in proceeding steps of data
preprocessing per omics type.}

\item{\code{Data_Checks}}{List of checks performed on omics data.}

\item{\code{Single_Omics}}{List of single omics analysis results per omics type.}

\item{\code{Multi_Omics}}{List of multi-omics integration results per
integration method.}

\item{\code{Ensemble}}{List of ensemble integration results.}

\item{\code{Plots}}{List of created plots.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ensemble_Integration_Suite-new}{\code{Ensemble_Integration_Suite$new()}}
\item \href{#method-Ensemble_Integration_Suite-add_Omics_file}{\code{Ensemble_Integration_Suite$add_Omics_file()}}
\item \href{#method-Ensemble_Integration_Suite-remove_Omics}{\code{Ensemble_Integration_Suite$remove_Omics()}}
\item \href{#method-Ensemble_Integration_Suite-add_Omics_df}{\code{Ensemble_Integration_Suite$add_Omics_df()}}
\item \href{#method-Ensemble_Integration_Suite-pivot_Samples_Wide}{\code{Ensemble_Integration_Suite$pivot_Samples_Wide()}}
\item \href{#method-Ensemble_Integration_Suite-run_Preprocessing}{\code{Ensemble_Integration_Suite$run_Preprocessing()}}
\item \href{#method-Ensemble_Integration_Suite-run_Sample_Matching}{\code{Ensemble_Integration_Suite$run_Sample_Matching()}}
\item \href{#method-Ensemble_Integration_Suite-run_Optimal_Cluster_Number}{\code{Ensemble_Integration_Suite$run_Optimal_Cluster_Number()}}
\item \href{#method-Ensemble_Integration_Suite-run_Single_Omics_Clustering}{\code{Ensemble_Integration_Suite$run_Single_Omics_Clustering()}}
\item \href{#method-Ensemble_Integration_Suite-run_Single_Omics_Hierarchical_Clustering}{\code{Ensemble_Integration_Suite$run_Single_Omics_Hierarchical_Clustering()}}
\item \href{#method-Ensemble_Integration_Suite-run_MoCluster}{\code{Ensemble_Integration_Suite$run_MoCluster()}}
\item \href{#method-Ensemble_Integration_Suite-run_MCIA}{\code{Ensemble_Integration_Suite$run_MCIA()}}
\item \href{#method-Ensemble_Integration_Suite-run_jNMF}{\code{Ensemble_Integration_Suite$run_jNMF()}}
\item \href{#method-Ensemble_Integration_Suite-run_iNMF}{\code{Ensemble_Integration_Suite$run_iNMF()}}
\item \href{#method-Ensemble_Integration_Suite-run_LRAcluster}{\code{Ensemble_Integration_Suite$run_LRAcluster()}}
\item \href{#method-Ensemble_Integration_Suite-run_COCA}{\code{Ensemble_Integration_Suite$run_COCA()}}
\item \href{#method-Ensemble_Integration_Suite-run_MOFA}{\code{Ensemble_Integration_Suite$run_MOFA()}}
\item \href{#method-Ensemble_Integration_Suite-run_SNF}{\code{Ensemble_Integration_Suite$run_SNF()}}
\item \href{#method-Ensemble_Integration_Suite-run_GAUDI}{\code{Ensemble_Integration_Suite$run_GAUDI()}}
\item \href{#method-Ensemble_Integration_Suite-run_Multi_Omics_Feature_HClust}{\code{Ensemble_Integration_Suite$run_Multi_Omics_Feature_HClust()}}
\item \href{#method-Ensemble_Integration_Suite-run_Ensemble_Sample_MDS}{\code{Ensemble_Integration_Suite$run_Ensemble_Sample_MDS()}}
\item \href{#method-Ensemble_Integration_Suite-run_Ensemble_Sample_CCA}{\code{Ensemble_Integration_Suite$run_Ensemble_Sample_CCA()}}
\item \href{#method-Ensemble_Integration_Suite-run_Ensemble_Integration_Feature}{\code{Ensemble_Integration_Suite$run_Ensemble_Integration_Feature()}}
\item \href{#method-Ensemble_Integration_Suite-run_Feature_Optimal_Cluster}{\code{Ensemble_Integration_Suite$run_Feature_Optimal_Cluster()}}
\item \href{#method-Ensemble_Integration_Suite-run_Factor_Optimal_Cluster}{\code{Ensemble_Integration_Suite$run_Factor_Optimal_Cluster()}}
\item \href{#method-Ensemble_Integration_Suite-run_Feature_Clustering}{\code{Ensemble_Integration_Suite$run_Feature_Clustering()}}
\item \href{#method-Ensemble_Integration_Suite-run_Factor_Clustering}{\code{Ensemble_Integration_Suite$run_Factor_Clustering()}}
\item \href{#method-Ensemble_Integration_Suite-run_Over_Represenation}{\code{Ensemble_Integration_Suite$run_Over_Represenation()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Heatmap_Cluster}{\code{Ensemble_Integration_Suite$plot_Heatmap_Cluster()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Multi_Omics_Heatmap}{\code{Ensemble_Integration_Suite$plot_Multi_Omics_Heatmap()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Scatterplot}{\code{Ensemble_Integration_Suite$plot_Scatterplot()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Clusters_Metadata}{\code{Ensemble_Integration_Suite$plot_Clusters_Metadata()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Sankey_Clusters}{\code{Ensemble_Integration_Suite$plot_Sankey_Clusters()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Sankey_Methods}{\code{Ensemble_Integration_Suite$plot_Sankey_Methods()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Top_Feature_Weights}{\code{Ensemble_Integration_Suite$plot_Top_Feature_Weights()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Dendrogram}{\code{Ensemble_Integration_Suite$plot_Dendrogram()}}
\item \href{#method-Ensemble_Integration_Suite-plot_Enrich_Dot}{\code{Ensemble_Integration_Suite$plot_Enrich_Dot()}}
\item \href{#method-Ensemble_Integration_Suite-clone}{\code{Ensemble_Integration_Suite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-new}{}}}
\subsection{Method \code{new()}}{
Create a new Ensemble_Integration_Suite object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new Ensemble_Integration_Suite object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-add_Omics_file"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-add_Omics_file}{}}}
\subsection{Method \code{add_Omics_file()}}{
Add individual omics raw data and metadata by  file path, to be
repeated for each omics dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$add_Omics_file(
  OmicsName,
  DataFile,
  MetadataFile,
  IndexColumn = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OmicsName}}{Name of omics type.}

\item{\code{DataFile}}{File path name for the raw omics data of the omics type.}

\item{\code{MetadataFile}}{File path name for the metadata of the omics type.}

\item{\code{IndexColumn}}{Whether to use the first column as the index. Default
is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-remove_Omics"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-remove_Omics}{}}}
\subsection{Method \code{remove_Omics()}}{
Remove all created datasets for a specific omics type. Removes both
the raw datafiles associated with a specific omics type, as well as
the preprocessed files. Analyses without this omics type will need to
be run again.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$remove_Omics(OmicsName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OmicsName}}{Name of the omics type to be removed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-add_Omics_df"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-add_Omics_df}{}}}
\subsection{Method \code{add_Omics_df()}}{
Add individual omics raw data and metadata from R environment, to be
repeated for each omics type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$add_Omics_df(OmicsName, DataDF, MetaDF)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OmicsName}}{Name of omics type.}

\item{\code{DataDF}}{Dataframe or matrix with raw omics data of the omics type.}

\item{\code{MetaDF}}{Dataframe or matrix with metadata of the omics type.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-pivot_Samples_Wide"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-pivot_Samples_Wide}{}}}
\subsection{Method \code{pivot_Samples_Wide()}}{
Transpose the omics matrices. Required if input raw omics data does
not store samples in columns and features in rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$pivot_Samples_Wide(SamplesInCol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SamplesInCol}}{Are samples in columns? If FALSE (default), the raw
omics matrices will be transposed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Preprocessing"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Preprocessing}{}}}
\subsection{Method \code{run_Preprocessing()}}{
Perform multiple different preprocessing and filtering steps on
individual raw omics datasets. Implemented steps and order can be
determined.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Preprocessing(
  OmicsName,
  FunctionOrder = c("FeatureProp", "FilterCoverage", "FilterBlankMean",
    "FilterCorrelation", "Normalization", "NAImpute", "BatchCorrection"),
  QCName = NULL,
  BlankName = NULL,
  PooledName = NULL,
  CoverageThreshold = 0.8,
  BlankMeanMultiplier = 2,
  BlankMeanThreshold = 0.8,
  CorrelationThreshold = 0.9,
  CorMean = FALSE,
  NAMethod = NULL,
  NA_K_Neighbors = 10,
  BatchColumnName = NULL,
  PreBatchPCA = TRUE,
  BatchPCAXAxis = 1,
  BatchPCAYAxis = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OmicsName}}{Name of omics type.}

\item{\code{FunctionOrder}}{Character vector with any of the following 7 steps
in any order.
\itemize{
\item \code{"FeatureProp"}: Calculate the feature proportion of feature total
per sample.
\item \code{"FilterCoverage"}: Filter features based on NA proportion across
samples, by user defined threshold.
\item \code{"FilterBlankMean"}: Filter features based on its measurement in
blank samples. Multiply the mean feature measurement of the blank
samples by a user defined multiplier, and filter the features in which
less than the user defined threshold exceed the multiplied blank mean.
Only possible if blank samples are available.
\item \code{"FilterCorrelation"}: Filter out correlated features. Calculate
pairwise correlations between features, and filter out the features
with correlation above the user defined threshold. One feature per
correlated group is kept in the dataset. The features which are
filtered out are stored in a list at
$Preprocessed_Omics$Correlated_Features.
\item \code{"Normalization"}: Feature Z-score normalization.
\item \code{"NAImpute"}: NA imputation or replacement with one of the following
methods: KNN imputation, missForest, zero replacement, omit NA.
\item \code{"BatchCorrection"}: Correct for batch effect using ComBat. Only
possible if batch information is available for samples.
}}

\item{\code{QCName}}{Common string in the name of quality control samples.}

\item{\code{BlankName}}{Common string in the name of blank samples.}

\item{\code{PooledName}}{Common string in the name of pooled samples.}

\item{\code{CoverageThreshold}}{Numerical threshold for feature coverage
filtering performed in \code{"FilterCoverage"}. Default is 0.8 (at least 0.8
of the samples do not have NAs in a feature).}

\item{\code{BlankMeanMultiplier}}{Numerical by which blank feature mean is
multiplied in \code{"FilterBlankMean"}. Default is 2.}

\item{\code{BlankMeanThreshold}}{Numerical threshold for blank mean filtering
performed in \code{"FilterBlankMean"}. Default is 0.8 (at least 0.8 of the
samples are larger than blank feature mean * BlankMeanMultiplier).}

\item{\code{CorrelationThreshold}}{Numerical threshold for correlation filtering
performed in \code{"FilterCorrelation"}. Default is 0.9 (if pairwise
correlation between two features is 0.9 or larger, one of the features is
removed).}

\item{\code{CorMean}}{Whether to calculate the mean measurements for the
correlated features. If \code{FALSE} (default), a representative feature from
the correlated group of features is kept in the dataset. If \code{TRUE}, the
mean of all feature measurements for a group of correlated features is
included in the dataset.}

\item{\code{NAMethod}}{String with method for \code{"NAImpute"}: \code{"KNN"},
\code{"missForest"}, \code{"zero replacement"} or \code{"omit NA"}.}

\item{\code{NA_K_Neighbors}}{Numerical for the number of K nearest neighbors
when using \code{"NAMethod"} \code{"KNN"}. Default is 10.}

\item{\code{BatchColumnName}}{String of the column name which contains batching
information in the metadata file.}

\item{\code{PreBatchPCA}}{Whether to create a PCA plot of the sample data prior
to batch effect correction. If \code{TRUE} (default), a PCA plot of the
samples before batch effect correction will be created and printed.}

\item{\code{BatchPCAXAxis}}{Numerical to indicate which principle component will
be used for X axis in the PCA plot. Default is 1.}

\item{\code{BatchPCAYAxis}}{Numerical to indicate which principle component will
be used for Y axis in the PCA plot. Default is 2.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Sample_Matching"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Sample_Matching}{}}}
\subsection{Method \code{run_Sample_Matching()}}{
Select the samples which match across the different omics datasets.
Most methods require the same samples in the same order in all
included omics datasets. This method also matches samples in the
metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Sample_Matching()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{E <- Ensemble_Integration_Suite$new()
E$add_Omics_df(OmicsName = "Proteomics",
DataDF = EINS_NCI60_Toy$Proteomics, MetaDF = EINS_NCI60_Toy$Metadata)
E$add_Omics_df(OmicsName = "Methylation",
DataDF = EINS_NCI60_Toy$Methylation,
MetaDF = EINS_NCI60_Toy$Methylation)=
E$run_Sample_Matching()
E$Preprocessed_Omics$Matched_Data$Proteomics

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Optimal_Cluster_Number"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Optimal_Cluster_Number}{}}}
\subsection{Method \code{run_Optimal_Cluster_Number()}}{
Calculate the optimal cluster number for integrated data using
NbClust. Performs data integration with seven methods that provide
sample by factor/component matrices: MoCluster, MCIA, jNMF, iNMF,
LRAcluster, MOFA and GAUDI. Integration methods are performed with
default parameters. A range of factors/components can be
provided, for multiple integration results to be used. NbClust is
then performed on these matrices. Multiple NbClust indexes are
calculated per matrix. All results from NbClust indexes for all
matrices are combined and the mean, mode and median optimal cluster
number are calculated. The user can then determine the preferred
cluster number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Optimal_Cluster_Number(
  Components = NULL,
  NbClustDistance = "euclidean",
  NbClustMinClust = 2,
  NbClustMaxClust = 12,
  NbClustMethod = "ward.D2",
  NbClustIndex = "all",
  NbClustAlphaBeale = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Components}}{Numerical or range of the components/factors to perform
the integrations with.}

\item{\code{NbClustDistance}}{String with the distance measure to be used for
the NbClust dissimilarity matrix computation. Must be one of:
\code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or
\code{"minkowski"}. Default is \code{"euclidean"}.}

\item{\code{NbClustMinClust}}{Minimal number of clusters considered by NbClust.
Value between 1 and (number of objects - 1). Default is 2.}

\item{\code{NbClustMaxClust}}{Maximum number of clusters considered by NbClust.
Value between 2 and (number of object -2), but larger than
NbClustMinClust. Default is 12.}

\item{\code{NbClustMethod}}{The cluster analysis method used by NbClust. Must be
one of: \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"},
\code{"mcquitty"}, \code{"median"}, \code{"centroid"}, \code{"kmeans"}. Default is
\code{"ward.D2"}.}

\item{\code{NbClustIndex}}{Indexes to be calculated by NbClust. Must be one of:
\code{"kl"}, \code{"ch"}, \code{"hartigan"}, \code{"ccc"}, \code{"scott"}, \code{"marriot"}, \code{"trcovw"},
\code{"tracew"}, \code{"friedman"}, \code{"rubin"}, \code{"cindex"}, \code{"db"}, \code{"silhouette"},
\code{"duda"}, \code{"pseudot2"}, \code{"beale"}, \code{"ratkowsky"}, \code{"ball"},
\code{"ptbiserial"}, \code{"gap"}, \code{"frey"}, \code{"mcclain"}, \code{"gamma"}, \code{"gplus"},
\code{"tau"}, \code{"dunn"}, \code{"hubert"}, \code{"sdindex"}, \code{"dindex"}, \code{"sdbw"}, \code{"all"}
(all indices except GAP, Gamma, Gplus and Tau), \code{"alllong"} (all indices
with Gap, Gamma, Gplus and Tau included). Default is \code{"all"}.}

\item{\code{NbClustAlphaBeale}}{Numerical, significance value for Beale's index.
Default is 0.1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Single_Omics_Clustering"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Single_Omics_Clustering}{}}}
\subsection{Method \code{run_Single_Omics_Clustering()}}{
Uses model-based clustering method HDDC from the HDclassif package
to provide clustering results for individual omics datasets.
This is a specific parametrization of the Gaussian mixture model to
work with high dimensional data. When running
\code{run_Single_Omics_Clustering}, clustering is performed on all omics
datasets imported into the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Single_Omics_Clustering(
  MinClust = 2,
  MaxClust = 12,
  HDModel = "AkjBkQkDk",
  HDThreshold = 0.2,
  HDCriterion = "BIC",
  HDItermax = 200,
  HDEps = 0.001,
  HDAlgo = "EM",
  HDD_select = "Cattell",
  HDInit = "kmeans",
  HDInit.vector,
  HDMini.nb = c(5, 10),
  HDScaling = FALSE,
  HDMin.individuals = 2,
  HDNoise.ctrl = 1e-08,
  HDMc.cores = 1,
  HDNb.rep = 1,
  HDKeepAllRes = TRUE,
  HDKmeans.control = list(),
  HDD_max = 100,
  HDSubset = Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MinClust}}{Minimum number of clusters considered by the model.
Default is 2.}

\item{\code{MaxClust}}{Maximum number of clusters considered by the model.
Value needs to be at least (MinClust + 1). Default is 12.}

\item{\code{HDModel}}{Name of the model. Must be one or combination of:
\code{"AkjBkQkDk"} (default), \code{"AkBkQkDk"}, \code{"ABkQkDk"}, \code{"AkjBQkDk"},
\code{"AkBQkDk"}, \code{"ABQkDk"}, \code{"AkjBkQkD"}, \code{"AkBkQkD"}, \code{"ABkQkD"},
\code{"AkjBQkD"}, \code{"AkBQkD"}, \code{"ABQkD"}, \code{"AjBQD"} or \code{"ABQD"}. If a
combination of models is selected, the model which maximizes the BIC
criterion is kept. All models can be run with \code{"ALL"}.}

\item{\code{HDThreshold}}{Threshold used in Cattell's Scree-Test. Must be
between 0 and 1. Default is 0.2.}

\item{\code{HDCriterion}}{Criterion used to select the best model. Either
\code{"BIC"} (default) or \code{"ICL"}.}

\item{\code{HDItermax}}{Maximum number of iteration allowed. Default is 200.}

\item{\code{HDEps}}{Double positive, default is 0.001. Used as the stopping
criterion: the algorithm stops if the difference between two successive
log-likelihoods is lower than \code{"HDEps"}.}

\item{\code{HDAlgo}}{The algorithm to be used. Must be one of: \code{"EM"}
(expectation-maximization), \code{"CEM"} (classification EM), or \code{"SEM"}
(stochastic EM). Default is \code{"EM"}.}

\item{\code{HDD_select}}{Method to select the intrinsic dimension. Must be one
of: \code{"Cattell"} (default), or \code{"BIC"}.}

\item{\code{HDInit}}{Method to initialize the EM algorithm. Must be one of:
\code{"kmeans"} (default), \code{"param"}, \code{"random"}, \code{"mini-em"} or \code{"vector"}.
If \code{"vector"} is selected, argument \code{"HDInit.vector"} should be added.}

\item{\code{HDInit.vector}}{Vector of integers or factors of the same length as
the number of samples. Only used if \code{"HDInit"} is \code{"vector"}.}

\item{\code{HDMini.nb}}{Vector of 2 integers. Parameter used in \code{"mini-em"}
initialization. First integer indicates the amount of times the algorithm
will be launched, the second integer indicates how many iterations are
performed each launch. Default is \code{"c(5,10)"}.}

\item{\code{HDScaling}}{Whether to perform scaling on the dataset. Default is
\code{FALSE}.}

\item{\code{HDMin.individuals}}{The minimum number of samples per cluster. Must
be 2 (default) or greater.}

\item{\code{HDNoise.ctrl}}{Parameter to minimize the intrinsic dimension
selected. Dimensions with eigenvalues lower than \code{"HDNoise.ctrl"} are not
used. Default is 1e-8.}

\item{\code{HDMc.cores}}{Whether parallel computing is used. Default is 1,
indicating no parallel computing. If \code{"HDMc.cores"} is larger than 1,
parallel computing is used with \code{"HDMc.cores"} cores.}

\item{\code{HDNb.rep}}{Number of times each estimation is repeated, with the
estimation with the highest log-likelihood kept. Default is 1 (no
repetition).}

\item{\code{HDKeepAllRes}}{Whether results of all runs should be kept. Default
is \code{TRUE}.}

\item{\code{HDKmeans.control}}{List of kmeans initialization parameters:
\code{"iter.max"}, \code{"nstart"} and \code{"algorithm"}.}

\item{\code{HDD_max}}{Maximum number of dimensions to be computed. Default is
100.}

\item{\code{HDSubset}}{Clustering can be performed on a subset of the data. If
\code{"HDSubset"} is smaller than the number of samples, clustering is
performed on a random subset of that size. Then, the posterior of the
clustering is computed on the full sample. Default is \code{Inf}, indicating
no subsetting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Single_Omics_Hierarchical_Clustering"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Single_Omics_Hierarchical_Clustering}{}}}
\subsection{Method \code{run_Single_Omics_Hierarchical_Clustering()}}{
Sample hierarchical clustering to be performed on each individual
omics dataset. When running \code{run_Single_Omics_Hierarchical_Clustering},
hierarchical clustering is performed on all omics datasets imported
into the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Single_Omics_Hierarchical_Clustering(
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Distance}}{Distance measure to be used. Must be one of:
\code{"euclidean"} (default), \code{"maximum"}, \code{"manhattan"}, \code{"canberra"},
\code{"binary"} or \code{"minkowski"}. If \code{"minkowski"} is selected, argument
\code{"MinkowskiPower"} needs to be included.}

\item{\code{MinkowskiPower}}{Power of the Minkowski distance. Default is NULL.}

\item{\code{Linkage}}{Agglomeration method to be used. Must be one of: \code{"ward.D"},
\code{"ward.D2"} (default), \code{"single"}, \code{"complete"}, \code{"average"}, \code{"mcquitty"},
\code{"median"} or \code{"centroid"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_MoCluster"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_MoCluster}{}}}
\subsection{Method \code{run_MoCluster()}}{
Perform multi-omics integration and sample clustering using MoCluster
from the mogsa package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_MoCluster(
  Clusters = NULL,
  Components = NULL,
  Method = "globalScore",
  Option = "lambda1",
  k = "all",
  Center = TRUE,
  Scale = FALSE,
  MaxIter = 1000,
  svdSolver = "fast.svd",
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "complete"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{The number of sample clusters to be created. Can be a
single integer to calculate a single clustering assignment, or a range to
calculate multiple clustering assignment.}

\item{\code{Components}}{The number of components to reduce the data to. Needs
be a single integer.}

\item{\code{Method}}{String to indicate the integration approach. Must be one of:
\code{"globalScore"} (consensus PCA), \code{"blockScore"} (generalized canonical
correlation analysis) or \code{"blockLoading"} (multiple co-inertia analysis).
Default is \code{"globalScore"}.}

\item{\code{Option}}{String to indicate how matrices should be normalized. Must
be one of: \code{"lambda1"} (matrix divided by first singular value),
\code{"inertia"} (matrix divided by its total inertia) or \code{"uniform"} (no
normalization). Default is \code{"lambda1"}.}

\item{\code{k}}{The number of non-zero coefficients for the variable loading
vectors to determine the sparsity. If \code{k} >= 1, \code{k} is the absolute
number of non-zero coefficients. If 0  < \code{k} < 1, \code{k} is the proportion
of non-zero coefficients. A single value or vector with length equal to
the number of omics datasets can be provided. Default is \code{"all"},
indicating no sparsity.}

\item{\code{Center}}{Whether the variables should be centered. Default is \code{TRUE}.}

\item{\code{Scale}}{Whether the variables should be scaled. Default is \code{FALSE}.}

\item{\code{MaxIter}}{Maximum number of iterations in the algorithm. Default is
1000.}

\item{\code{svdSolver}}{String to indicate the method used for singular value
decomposition. Must be one of: \code{"fast.svd"} (default), \code{"svd"} or
\code{"propack"}. \code{"fast.svd"} has a good compromise between speed and
robustness compared to \code{"svd"} (slower but more robust) and \code{"propack"}
(faster but less likely to converge).}

\item{\code{Distance}}{Distance measure to be used for distance matrix
calculation of the integrated data. Must be one of: \code{"euclidean"}
(default), \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or
\code{"minkowski"}. If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"}
needs to be included.}

\item{\code{MinkowskiPower}}{Power of the Minkowski distance. Default is NULL.}

\item{\code{Linkage}}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"}, \code{"single"},
\code{"complete"} (default), \code{"average"}, \code{"mcquitty"}, \code{"median"} or
\code{"centroid"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_MCIA"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_MCIA}{}}}
\subsection{Method \code{run_MCIA()}}{
Perform multi-omics integration using multiple co-inertia analysis
(MCIA) from the omicade4 package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_MCIA(
  PerformCluster = TRUE,
  Clusters = NULL,
  Components = NULL,
  SVD = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PerformCluster}}{Whether to perform k-means clustering on the sample
embeddings. Default is \code{TRUE}.}

\item{\code{Clusters}}{The number of sample clusters to be created. Can be a
single integer to calculate a single cluster assignment, or a range to
calculate multiple cluster assignments.}

\item{\code{Components}}{The number of components to reduce the data to. Needs
to be a single integer.}

\item{\code{SVD}}{Whether singular value decomposition should be used. Default
is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_jNMF"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_jNMF}{}}}
\subsection{Method \code{run_jNMF()}}{
Perform multi-omics integration using joint non-negative matrix
factorization (jNMF) from the nnTensor package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_jNMF(
  Clusters = NULL,
  NMFAlgorithm = "KL",
  MaxIter = 100
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{The number of sample clusters to be created. Can be a
single integer to calculate a single cluster assignment, or a range to
calculate multiple cluster assignments. In jNMF, the number of clusters
is the number of low dimensions or components created.}

\item{\code{NMFAlgorithm}}{String indicating algorithm for the divergence
between X and X-bar. Must be one of: \code{"KL"} (default), \code{"Frobenius"} or
\code{"IS"}.}

\item{\code{MaxIter}}{Number of iterations. Default is 100.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_iNMF"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_iNMF}{}}}
\subsection{Method \code{run_iNMF()}}{
Perform multi-omics integration and sample clustering using
integrative non-negative matrix factorization from the IntNMF package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_iNMF(
  Clusters = NULL,
  MaxIter = 200,
  StabilityCount = 20,
  MatrixInit = 30
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{The number of sample clusters to be created. Can be a
single integer to calculate a single cluster assignment, or a range to
calculate multiple cluster assignments. In iNMF, the number of clusters
is the number of low dimensions or components created.}

\item{\code{MaxIter}}{Number of iterations. Default is 200.}

\item{\code{StabilityCount}}{Count for stability in the connectivity matrix.
Default is 20.}

\item{\code{MatrixInit}}{Number of initializations of random matrices. Default
is 30.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_LRAcluster"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_LRAcluster}{}}}
\subsection{Method \code{run_LRAcluster()}}{
Perform multi-omics integration and sample clustering using
LRAcluster for R as done in the MOVICS package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_LRAcluster(Clusters = NULL, Linkage = "ward.D")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{The Number of sample clusters to be created. Can be a
single integer to calculate a single cluster assignment, or a range to
calculate multiple cluster assignments.}

\item{\code{Linkage}}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"} (default), \code{"ward.D2"}, \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_COCA"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_COCA}{}}}
\subsection{Method \code{run_COCA()}}{
Perform multi-omics integration and sample clustering using cluster
of cluster analysis from the COCA package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_COCA(
  DatasetClusters = NULL,
  maxClusters = 12,
  DatasetClusterMethod = "hclust",
  DatasetDistance = "euclidean",
  FullClusters = NULL,
  maxFullClusters = 12,
  FullHClustMethod = "average",
  FullClusterMethod = "kmeans",
  FullDistance = "euclidean"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetClusters}}{Number of clusters per omics dataset. If a single
integer, all omics datasets get the same number of clusters. If a vector
with the same length as the number of omics datasets, each dataset can get
a different number of clusters. If NULL, the number of clusters per
dataset will be estimated using the silhouette method, and argument
\verb{"maxClusters} should be included.}

\item{\code{maxClusters}}{Maximum number of clusters per omics dataset if
\code{DatasetClusters} is NULL. If a single integer, all omics datasets have the
same maximum number of clusters considered. If a vector with the same
length as the number of omics datasets, each dataset has an individual
maximum number of clusters considered. Default is 12.}

\item{\code{DatasetClusterMethod}}{String indicating the method to cluster the
individual omics datasets. If single string, all omics datasets are
clustered with the same method. If vector of strings, each omics dataset
is clustered with an individual method. Must be one of: \code{"kmeans"},
\code{"hclust"} (default) or \code{"pam"}.}

\item{\code{DatasetDistance}}{Distance to be used in clustering of the
individual omics datasets. If single string, all omics datasets are
clustered with the same distance. If vector of strings, each omics
dataset is clustered with an individual distances. Must be one of:
\code{"euclidean"} (default, available for \code{"kmeans"}, \code{"hclust"} and \code{"pam"}),
\code{"manhattan"} (available for \code{"kmeans"}, \code{"hclust"} and \code{"pam"}),
\code{"gower"} (available for \code{"pam"}), \code{"maximum"} (available for \code{"kmeans"}
and \code{"hclust"}), \code{"canberra"} (available for \code{"kmeans"} and \code{"hclust"}),
\code{"binary"} (available for \code{"kmeans"} and \code{"hclust"}) or \code{"minkowski"}
(available for \code{"kmeans"} and \code{"hclust"}).}

\item{\code{FullClusters}}{Number of clusters after integration. Can be a single
integer to calculate a single cluster assignment, or a range to calculate
multiple cluster assignments. If NULL, the number of clusters will be
estimated using the silhouette method, and argument \code{"maxFullClusters"}
should be included.}

\item{\code{maxFullClusters}}{Maximum number of clusters if \code{FullClusters} is
NULL. Default is 12.}

\item{\code{FullHClustMethod}}{Agglomeration method to be used for the
hierarchical clustering of the integrated dataset. Must be one of:
\code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, "complete"\verb{, }"average"\verb{(default),}"mcquitty"\verb{, }"median"\code{ or}"centroid"`.}

\item{\code{FullClusterMethod}}{String indicating the method to cluster the
integrated dataset. Must be one of: \code{"kmeans"} (default) or \code{"hclust"}.}

\item{\code{FullDistance}}{Distance to be used in clustering of the
integrated dataset. Must be one of: \code{"pearson"}, \code{"spearman"},
\code{"euclidean"} (default), \code{"manhattan"}, \code{"maximum"}, \code{"canberra"},
\code{"binary"} or \code{"minkowski"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_MOFA"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_MOFA}{}}}
\subsection{Method \code{run_MOFA()}}{
Perform multi-omics integration using multi-omics factor analysis
from the MOFA2 package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_MOFA(
  PerformCluster = TRUE,
  Clusters = NULL,
  Factors = NULL,
  ScaleViews = FALSE,
  SpikeSlabFactors = FALSE,
  SpikeSlabWeights = TRUE,
  ARDFactors = FALSE,
  ARDWeights = TRUE,
  MaxIter = 1000,
  Convergence = "fast",
  StartELBO = 1,
  FreqELBO = 1,
  Stochastic = FALSE,
  SaveData = FALSE,
  Outfile = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PerformCluster}}{Whether to perform k-means clustering on the sample
embeddings. Default is \code{TRUE}.}

\item{\code{Clusters}}{The number of sample clusters to be created. Can be a
single integer to calculate a single cluster assignment, or a range to
calculate multiple cluster assignments.}

\item{\code{Factors}}{Number of factors (components).}

\item{\code{ScaleViews}}{Whether to scale the different omics datasets to be
scaled to have the same unit variance. Default is \code{FALSE}.}

\item{\code{SpikeSlabFactors}}{Whether to use spike and slab sparsity on the
factors (components). Default is \code{FALSE}.}

\item{\code{SpikeSlabWeights}}{Whether to use spike and slab sparsity on the
features. Default is \code{TRUE}.}

\item{\code{ARDFactors}}{Whether to use ARD sparsity on the factors. Default is
\code{FALSE}.}

\item{\code{ARDWeights}}{Whether to use ARD sparsity on the weights. Default is
\code{TRUE}.}

\item{\code{MaxIter}}{Maximum number of iterations. Default is 1000.}

\item{\code{Convergence}}{String indicating the convergence criteria. Must be
one of: \code{"fast"} (default), \code{"medium"}, or \code{"slow"}.}

\item{\code{StartELBO}}{Integer indicating the first iteration to compute the
ELBO statistic. Default is 1.}

\item{\code{FreqELBO}}{Integer indicating the frequency of ELBO computations.
Default is 1.}

\item{\code{Stochastic}}{Whether to use stochastic inference. Default is \code{FALSE}.}

\item{\code{SaveData}}{Whether to save the trained MOFA model. Default is
\code{FALSE}.}

\item{\code{Outfile}}{File path to save the trained MOFA model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_SNF"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_SNF}{}}}
\subsection{Method \code{run_SNF()}}{
Perform multi-omics integration and sample clustering using
similarity network fusion from the SNFtool package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_SNF(
  Distance = "euclidean squared",
  Clusters = NULL,
  Neighbors = 20,
  Sigma = 0.5,
  Iter = 10,
  SpecClusterType = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Distance}}{String indicating the distance used to calculate the
distance matrices for each omics dataset. If single string, distance
matrices are only calculated for that distance. If a vector of strings,
distance matrices are calculated for all distances, and clustering will
be performed for all distances. Can be: \code{"euclidean squared"} (default
in SNF), \code{"euclidean"}, \code{"manhattan"}, \code{"minkowski 0.25"},
\code{"minkowski 0.5"}, \code{"minkowski 3"}, \code{"minkowski 4"}.}

\item{\code{Clusters}}{Number of sample clusters to be created. Can be a single
integer to calculate a single cluster assignment, or a range to calculate
multiple cluster assignments.}

\item{\code{Neighbors}}{Number of nearest neighbors. Default is 20.}

\item{\code{Sigma}}{Variance for the local model. Default is 0.5.}

\item{\code{Iter}}{Number of iterations for the diffusion process. Default is 10.}

\item{\code{SpecClusterType}}{Variant of spectral clustering to use. Must be one
of: 1, 2 or 3 (default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_GAUDI"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_GAUDI}{}}}
\subsection{Method \code{run_GAUDI()}}{
Multi-omics integration and sample clustering using UMAP dimensionality
reduction as described in the GAUDI method. In addition to clustering
with HDBSCAN, as done in GAUDI, we provide DBSCAN and k-means
methods for sample clustering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_GAUDI(
  SingleOmicsUMAPNeighbors = 15,
  SingleOmicsUMAPDimensions = 4,
  SingleOmicsUMAPDistance = "euclidean",
  SingleOmicsUMAPScale = FALSE,
  SingleOmicsUMAPInit = "spectral",
  SingleOmicsUMAPPCA = NULL,
  ConcatUMAPNeighbors = 15,
  ConcatUMAPDimensions = 4,
  ConcatUMAPDistance = "euclidean",
  ConcatUMAPScale = FALSE,
  ConcatUMAPInit = "spectral",
  ConcatUMAPPCA = NULL,
  ClusterMethod = "HDBSCAN",
  HDBSCANminPts = 2,
  DBSCANeps = NULL,
  DBSCANminPts = 2,
  KmeansClusters = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SingleOmicsUMAPNeighbors}}{Size of the local neighborhood used for
manifold approximation for each individual omics dataset. If a single
integer, all omics datasets get the same size. If a vector with the same
length as the number of omics datasets, each omics dataset gets a
different size. Default is 15, recommended range is 2-100.}

\item{\code{SingleOmicsUMAPDimensions}}{The number of dimensions into which the
individual omics datasets are embedded. If a single integer, all omics
datasets get the same number of dimensions. If a vector with the same
length as the number of omics datasets, each omics dataset gets a
different number of dimensions. Default is 4.}

\item{\code{SingleOmicsUMAPDistance}}{The distance metric used to find the
nearest neighbors in the individual omics datasets. If a single string,
all omics datasets use the same distance metric. If a vector of strings
with the same length as the number of omics datasets, each omics dataset
uses a different distance metric. Must be one of: \code{"euclidean"}
(default), \code{"cosine"}, \code{"manhattan"}, \code{"hamming"}, \code{"correlation"} or
\code{"categorical"}.}

\item{\code{SingleOmicsUMAPScale}}{Whether to perform scaling on the features of
the individual omics datasets. If a single boolean, all omics datasets
use the same scaling setting. If a vector of booleans, each omics dataset
uses a different scaling setting. Default is \code{FALSE}.}

\item{\code{SingleOmicsUMAPInit}}{Type of initialization for the coordinates of
the individual omics datasets. If a single string, all omics datasets use
the same initialization type. If a vector of strings, each omics dataset
uses a different initialization type. Must be one of: \code{"spectral"}
(default), \code{"normlaplacian"}, \code{"random"}, \code{"lvrandom"}, \code{"laplacian"},
\code{"pca"}, \code{"spca"} or \code{"agspectral"}.}

\item{\code{SingleOmicsUMAPPCA}}{Whether to reduce the number of features in the
individual omics datasets using PCA to increase the performance. May come
at a cost to accuracy. A positive integer less than the number of
features in the omics dataste is required. If a single integer, all omics
datasets are reduced to the same number of principle components. If a
vector of integers, each omics dataset is reduced to a different number
of principle components. Default is \code{NULL}.}

\item{\code{ConcatUMAPNeighbors}}{Size of the local neighborhood used for
manifold approximation for the UMAP calculation of the concatenated
single omics UMAP embeddings. Default is 15, recommended range is 2-100.}

\item{\code{ConcatUMAPDimensions}}{The number of dimensions into which the
concatenated single omics UMAP embeddings are embedded. Default is 4.}

\item{\code{ConcatUMAPDistance}}{The distance metric used to find the nearest
neighbors for the concatenated single omics UMAP embeddings. Must be one
of: \code{"euclidean"} (default), \code{"cosine"}, \code{"manhattan"}, \code{"hamming"},
\code{"correlation"} or \code{"categorical"}.}

\item{\code{ConcatUMAPScale}}{Whether to perform scaling on the UMAP dimensions
of the concatenated single omics UMAP embeddings. Default is \code{FALSE}.}

\item{\code{ConcatUMAPInit}}{Type of initialization for the coordinates of the
concatenated single omics UMAP embeddings. Must be one of: \code{"spectral"}
(default), \code{"nornlaplacian"}, \code{"random"}, \code{"lvrandom"}, \code{"laplacian"},
\code{"pca"}, \code{"spca"} or \code{"agspectral"}.}

\item{\code{ConcatUMAPPCA}}{Whether to reduce the number of UMAP dimensions of
the concatenated single omics UMAP embeddings using PCA to increase the
performance. May come at a cost to accuracy. Requires a positive integer
less than the the total number of UMAP dimension in all single omics UMAP
embeddings. Default is \code{NULL}.}

\item{\code{ClusterMethod}}{Method used for clustering of the concatenated UMAP
embedding. Must be one of: \code{"HDBSCAN"}, \code{"DBSCAN"} or \code{"Kmeans"}.}

\item{\code{HDBSCANminPts}}{Integer of the minimal size of the clusters. Default
is 2. Only included if \code{ClusterMethod} = \code{"HDBSCAN"}.}

\item{\code{DBSCANeps}}{Radius of the epsilon neighborhood. Default is \code{NULL}.
Only included if \code{ClusterMethod} = \code{"DBSCAN"}.}

\item{\code{DBSCANminPts}}{Integer of the minimal size of the clusters. Default
is 2. Only included if \code{ClusterMethod} = \code{"DBSCAN"}.}

\item{\code{KmeansClusters}}{Number of sample clusters to be created. Can be a
single integer to calculate a single cluster assignment, or a range to
calculate multiple cluster assignments. Only included if \code{ClusterMethod}
= \code{"Kmeans"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Multi_Omics_Feature_HClust"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Multi_Omics_Feature_HClust}{}}}
\subsection{Method \code{run_Multi_Omics_Feature_HClust()}}{
Hierarchical clustering for the feature weights calculated by the
multi-omics integration methods MoCluster, MCIA, jNMF, iNMF and MOFA.
Integrated omics feature datasets from $Multi_Omics$FeatureRes are
scaled and concatenated, and distance matrices are calculated per
multi-omics integration method. Hierarchical clustering is then
performed to cluster features from different omics types.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Multi_Omics_Feature_HClust(
  nFactors = NULL,
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nFactors}}{Number of factors (components) for which the feature
hierarchical clustering is to be performed. The multi-omics integration
methods need to be performed with this number of components (MoCluster
and MCIA), clusters (jNMF and iNMF) or factors (MOFA).}

\item{\code{Distance}}{Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{\code{MinkowskiPower}}{Power of the Minkowski distance. Default is NULL.}

\item{\code{Linkage}}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"} (default), \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Ensemble_Sample_MDS"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Ensemble_Sample_MDS}{}}}
\subsection{Method \code{run_Ensemble_Sample_MDS()}}{
Create an ensemble clustering result by combining the clustering
results from MoCluster, MCIA, jNMF, iNMF, LRAcluster, COCA, MOFA,
SNF and GAUDI. Borrows the idea of consensus cluster, by creating a
sample similarity matrix counting the number of times samples cluster
together. Hierarchical clustering is then performed on this matrix.
Provides a more robust clustering result when performed with
multiple cluster assignment, which can be achieved by running the
multi-omics integration methods with a range of \code{Clusters}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Ensemble_Sample_MDS(
  Clusters = NULL,
  SNFDistance = "euclidean squared",
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{Number of clusters to create from the hierarchical
clustering result.}

\item{\code{SNFDistance}}{Which distance used to calculate SNF results to be
included in the ensemble clustering. This is needed to reduce the impact
of SNF on the ensemble clustering assignment. Must be one of:
\code{"euclidean squared"} (default), \code{"euclidean"}, \code{"manhattan"},
\code{"minkowksi 0.25"}, \code{"minkowski 0.5"}, \code{"minkowski 3"} or \code{"minkowski 4"}.}

\item{\code{Distance}}{Distance Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{\code{MinkowskiPower}}{Power of the Minkowski distance. Default is NULL.}

\item{\code{Linkage}}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"}, \code{"single"},
\code{"complete"} (default), \code{"average"}, \code{"mcquitty"}, \code{"median"} or
\code{"centroid"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Ensemble_Sample_CCA"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Ensemble_Sample_CCA}{}}}
\subsection{Method \code{run_Ensemble_Sample_CCA()}}{
Create an ensemble clustering result by aligning the embeddings results
from MoCluster, MCIA, jNMF, iNMF, LRAcluster, COCA, MOFA, SNF and GAUDI.
Uses CCA to align the embeddings from the multi-omics integration methods
into a common latend space. For COCA and SNF, low-dimensional embeddings
of samples are not provided, and one-hot encoding is used in place of
embedding. One method is selected as reference embedding, with all other
embeddings sequentially aligned to this reference using pairwise CCA. The
aligned embedding is then used for k-means clustering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Ensemble_Sample_CCA(
  Clusters = NULL,
  SNFDistance = "euclidean squared",
  Reference = "MoCluster",
  Factors = NULL,
  GAUDIMethod = "Kmeans"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{Number of clusters to create from the CCA aligned
embedding.}

\item{\code{SNFDistance}}{Which distance used to calculate SNF results to be
included in the ensemble clustering. This is needed to reduce the impact
of SNF on the ensemble clustering assignment. Must be one of:
\code{"euclidean squared"} (default), \code{"euclidean"}, \code{"manhattan"},
\code{"minkowksi 0.25"}, \code{"minkowski 0.5"}, \code{"minkowski 3"} or \code{"minkowski 4"}.}

\item{\code{Reference}}{Multi-omics integration embedding to be used as
reference for CCA alignment.}

\item{\code{Factors}}{Number of factors for which the multi-omics integration
embedding is to be selected.}

\item{\code{GAUDIMethod}}{Which clustering method is used to calculate GAUDI
results to be included in the ensemble clustering. This is needed to
reduce the impact of GAUDI on the ensemble clustering assignment. Must be
one of: \code{"Kmeans"} (default), \code{"HDBSCAN"} or \code{"DBSCAN"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Ensemble_Integration_Feature"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Ensemble_Integration_Feature}{}}}
\subsection{Method \code{run_Ensemble_Integration_Feature()}}{
Create an ensemble feature clustering result by combining the feature
weight results from MoCluster, MCIA, jNMF, iNMF and MOFA. Feature
weigths must be calculated for one number of components (MoCluster
and MICA), clusters (jNMF and iNMF) and factors (MOFA) to be included.
Two methods of combining the feature weights are available:
\code{"Average"} and \code{"Concatenation"}. More details of the methods are
available in the \code{param} \code{Method}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Ensemble_Integration_Feature(
  Method = "Average",
  nFactors = NULL,
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Method}}{Method used for combining the feature weight matrices of
the different multi-omics integration methods. Must be one of:
\itemize{
\item \code{"Average"}: Feature weight matrices of the different omics types
are combined per multi-omics integration method. Feature distances are
calculated per method. The average feature distance between the five
methods is calculated. Hierachical clustering is performed on the
average feature distance matrix. Default method.
\item \code{"Concatenation"}: Feature weight matrices of the different omics
types are combined per multi-omics intergration method, which are then
concatenated into a large matrix. A feature distance matrix as well as
a factor distance matrix is calculated. Hierarchical clustering is
then performed for both the features and the factors per method.
}}

\item{\code{nFactors}}{Number of factors (components) for which the feature
hierarchical clustering is to be performed. The multi-omics integration
methods need to be performed with this number of components (MoCluster
and MCIA), clusters (jNMF and iNMF) or factors (MOFA).}

\item{\code{Distance}}{Distance Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{\code{MinkowskiPower}}{Power of the Minkowski distance. Default is NULL.}

\item{\code{Linkage}}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"}, \code{"single"},
\code{"complete"} (default), \code{"average"}, \code{"mcquitty"}, \code{"median"} or
\code{"centroid"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Feature_Optimal_Cluster"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Feature_Optimal_Cluster}{}}}
\subsection{Method \code{run_Feature_Optimal_Cluster()}}{
Calculate the optimal number of feature clusters using the feature
hierarchical clustering tree calculated in
\code{run_Feature_Ensemble_HClust}. Will calculate the optimal number of
feature clusters for both \code{"Average"} and \code{"Concatenation"} methods
at the same time. Uses the as.clustrange function from the
WeightedCluster package to calculate the optimal number of clusters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Feature_Optimal_Cluster(
  WeightedClusterStat = "all",
  MaxClusters = 20
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{WeightedClusterStat}}{The statistics included to calculate the
optimal cluster number. Must be one of: \code{"all"} (default), \code{"noCH"} (all
statistics except \code{"CH"} and \code{"CHsq"}), \code{"PBC"}, \code{"HG"}, \code{"HGSD"},
\code{"ASW"}, \code{"ASWw"}, \code{"CH"}, \code{"R2"}, \code{"CHsq"}, \code{"R2sq"} or \code{"HC"}.}

\item{\code{MaxClusters}}{The maximum number of feature clusters to be
considered by the statistics. Default is 20.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Factor_Optimal_Cluster"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Factor_Optimal_Cluster}{}}}
\subsection{Method \code{run_Factor_Optimal_Cluster()}}{
Calculate the optimal number of factor clusters using the factor
hierarchical clustering tree calculated in
\code{run_Feature_Ensemble_HClust} with \code{Method} \code{"Concatenation"}. Uses the
as.clustrange function from the WeightedCluster package to calculate the
optimal number of clusters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Factor_Optimal_Cluster(
  WeightedClusterStat = "all",
  MaxClusters = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{WeightedClusterStat}}{The statistics included to calculate the
optimal cluster number. Must be one of: \code{"all"} (default), \code{"noCH"} (all
statistics except \code{"CH"} and \code{"CHsq"}), \code{"PBC"}, \code{"HG"}, \code{"HGSD"},
\code{"ASW"}, \code{"ASWw"}, \code{"CH"}, \code{"R2"}, \code{"CHsq"}, \code{"R2sq"} or \code{"HC"}.}

\item{\code{MaxClusters}}{The maximum number of factor clusters to be
considered by the statistics. Default is 10.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Feature_Clustering"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Feature_Clustering}{}}}
\subsection{Method \code{run_Feature_Clustering()}}{
Cut the feature hierarchical clustering tree to create feature
clusters. Can be done either using the calculated optimal feature
cluster numbers, or with user argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Feature_Clustering(
  Clusters = NULL,
  Method = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{Number of clusters in which to cut the feature
hierarchical clustering tree. Default is \code{NULL.} If \code{NULL}, the
optimal feature cluster numbers calculated with
\code{run_Feature_Optimal_Cluster} are used, and hierarchical clustering trees
for all available \code{run_Feature_Ensemble_HClust} \code{Method} results. If an
integer, the \code{run_Feature_Ensemble_HClust} \code{Method} whose hierarchical
clustering tree is to be cut must be provided in parameter \code{Method}.}

\item{\code{Method}}{Which hierarchical clustering tree must be cut. Only
required if \code{Clusters} is an integer. Must be one of: \code{"Average"} or
\code{"Concatenation"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Factor_Clustering"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Factor_Clustering}{}}}
\subsection{Method \code{run_Factor_Clustering()}}{
Cut the factor hierarchical clustering tree to create factor
clusters. Can be done either using the calculated optimal factor
cluster numbers, or with user argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Factor_Clustering(Clusters = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clusters}}{Number of clusters in which to cut the factor
hierarchical clustering tree. Default is \code{NULL.} If \code{NULL}, the
optimal factor cluster numbers calculated with
\code{run_Factor_Optimal_Cluster} are used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-run_Over_Represenation"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-run_Over_Represenation}{}}}
\subsection{Method \code{run_Over_Represenation()}}{
Perform overrepresentation analysis from package ClusterProfiler
on the clustered ensembled features to determine if features with
similar biological functions group together in across multi-omics
integration methods. Uses enrichment GO categories from the genome
wide annotation database for human as provided in package
org.Hs.eg.db. Requires ensemble feature hierarchical clustering to
have been performed with \code{run_Feature_Ensemble_HClust} and a cluster
assignment to have been created with \code{run_Feature_Clustering}. Only
available for gene-based omics types (genomics, transcriptomics,
methylation etc.).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$run_Over_Represenation(
  Method = "Average",
  OmicsName,
  GeneNameType,
  Subontologies = "ALL",
  pValue = 0.05,
  pAdjustment = "BH",
  qValue = 0.2,
  MinGenesPerTerm = 10,
  MaxGenesPerTerm = 500
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Method}}{Which ensemble feature clustering result is to be used.
Must be one of" \code{"Average"} or \code{"Concatenation"}. Default is \code{"Average"}.}

\item{\code{OmicsName}}{For which omics dataset the overrepresentation analysis
is to be performed. Only genetics-based omics types can be utilized in the
genome wide annotation database.}

\item{\code{GeneNameType}}{Keytype of the feature name for the omics dataset.
Must be one of: \code{"ENTREZID"}, \code{"PFAM"}, \code{"IPI"}, \code{"PROSITE"}, \code{"ACCNUM"},
\code{"ALIAS"}, \code{"CHR"}, \code{"CHRLOC"}, \code{"CHRLOCEND"}, \code{"ENZYME"}, \code{"MAP"},
\code{"PATH"}, \code{"PMID"}, \code{"REFSEQ"}, \code{"SYMBOL"}, \code{"UNIGENE"}, \code{"ENSEMBL"},
\code{"ENSEMBLPROT"}, \code{"ENSEMBLTRANS"}, \code{"GENENAME"}, \code{"UNIPROT"}, \code{"GO"},
\code{"EVIDENCE"}, \code{"ONTOLOGY"}, \code{"GOALL"}, \code{"EVIDENCEALL"}, \code{"ONTOLOGYALL"},
\code{"OMIM"} or \code{"UCSCKG"}.}

\item{\code{Subontologies}}{Which ontology to use. Must be one of: \code{"ALL"}
(default), \code{"MF"}, \code{"BP"} or \code{"CC"}.}

\item{\code{pValue}}{Cutoff value for the p-value. Default is 0.05.}

\item{\code{pAdjustment}}{Method for the p-value adjustment. Must be one of:
\code{"BH"} (default), \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"},
\code{"BY"}, \code{"fdr"} or \code{"none"}.}

\item{\code{qValue}}{Cutoff for the q-value. Default is 0.2.}

\item{\code{MinGenesPerTerm}}{Minimal number of genes annotated per ontology
term to be included for testing. Default is 10.}

\item{\code{MaxGenesPerTerm}}{Maximum number of genes annotated per ontology
term to be included for testing. Default is 500.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Heatmap_Cluster"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Heatmap_Cluster}{}}}
\subsection{Method \code{plot_Heatmap_Cluster()}}{
Interactive heatmap with sample clustering results per omics type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Heatmap_Cluster(
  FeatureClustering = TRUE,
  MethodResults = "all",
  YFontSize = 5,
  XFontSize = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureClustering}}{Whether to order the features on the Y-axis
based on hierarchical clustering. Default is \code{TRUE}. If \code{FALSE}, the
features will be ordered as in the preprocessed datasets.}

\item{\code{MethodResults}}{For which clustering assignment the heatmap is to be
plotted. Must be one of: \code{"all"} (default), \code{"MoCluster"}, \code{"jNMF"}, \code{"iNMF"},
\code{"LRAcluster"}, \code{"COCA"}, \code{"GAUDI"}, \code{"SNF"}, \code{"MDS"} or \code{"CCA"}. If all is
selected, heatmaps with cluster assignments are created for all run
methods.}

\item{\code{YFontSize}}{Font size of the feature labels on the Y-axis. Default
is 5.}

\item{\code{XFontSize}}{Font size of the sample labels on the X-axis. Default
is 5.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Multi_Omics_Heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Multi_Omics_Heatmap}{}}}
\subsection{Method \code{plot_Multi_Omics_Heatmap()}}{
Non-interactive heatmap with all omics heatmaps in a single plot as
well as clustering results and metadata features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Multi_Omics_Heatmap(
  LegendNames = NULL,
  MetadataFeatures = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LegendNames}}{Names for the legends of the different omics heatmaps.
If \code{NULL}, the omics names as used in the EINS object are used. Otherwise,
a string vector with the same length as number of omics datasets should be
provided.}

\item{\code{MetadataFeatures}}{Names of the sample metadata features to be
displayed on the heatmaps. These metadata features need to be available
in the metadata files for all omics types. A string vector with the names
of the metadata features as in the metadata files should be provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Scatterplot"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Scatterplot}{}}}
\subsection{Method \code{plot_Scatterplot()}}{
Interactive scatterplot of samples in the factors
(components/dimensions) for methods which have CoordData (MoCluster,
MCIA, LRAcluster, MOFA and GAUDI).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Scatterplot(
  XAxis = 1,
  YAxis = 2,
  YFontSize = 15,
  XFontSize = 15,
  Width = 5,
  Height = 5,
  MetadataFeature = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{XAxis}}{Integer, factor (component/dimension) to plot on the X-axis.
Default is 1.}

\item{\code{YAxis}}{Integer, factor (component/dimension) to plot on the Y-axis.
Default is 2.}

\item{\code{YFontSize}}{Size of the Y-axis title. Default is 15.}

\item{\code{XFontSize}}{Size of the X-axis title. Default is 15.}

\item{\code{Width}}{Width of the scatterplot. Default is 5.}

\item{\code{Height}}{Height of the scatterplot. Default is 5.}

\item{\code{MetadataFeature}}{Name of the metadata feature used as the shape of
the dots. Must be a string with the name of the metadata feature as in
the metadata files. This metadata feature needs to be available in the
metadatafiles of all omics types.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Clusters_Metadata"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Clusters_Metadata}{}}}
\subsection{Method \code{plot_Clusters_Metadata()}}{
Table with sample cluster assignment and metadata features, with each
column colored by level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Clusters_Metadata(MetadataFeatures = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MetadataFeatures}}{Names of the sample metadata features to be
included in the table. These metadata features need to be available
in the metadata files for all omics types. A string vector with the names
of the metadata features as in the metadata files should be provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Sankey_Clusters"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Sankey_Clusters}{}}}
\subsection{Method \code{plot_Sankey_Clusters()}}{
Interactive Sankey plot of the sample cluster results per method.
Compares the cluster assignments of a single method with multiple
cluster numbers. Can only be created if multiple cluster numbers are
calculated per multi-omics integration method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Sankey_Clusters(MetadataFeature = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MetadataFeature}}{String of the sample metadata feature for which the
samples should be colored. The metadata feature needs to be available
in the metadata files for all omics types.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Sankey_Methods"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Sankey_Methods}{}}}
\subsection{Method \code{plot_Sankey_Methods()}}{
Interactive Sankey plot of sample cluster results between methods.
Compares the cluster assignments of all multi-omics integration
methods which have the same number of clusters. Can only be performed
if multiple integration methods have calculated cluster assignments
for the same cluster number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Sankey_Methods(
  MetadataFeature = NULL,
  Ensemble = TRUE,
  SNFDistance = "euclidean squared",
  AllPlots = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MetadataFeature}}{String of the sample metadata feature for which the
samples should be colored. The metadata feature need to be available
in the metadata files for all omics types.}

\item{\code{Ensemble}}{Whether to include the ensemble sample clustering results
to the Sankey plot. Must run \code{run_Ensemble_Sample_MDS} or
\code{run_Ensemble_Sample_CCA} first. Default is \code{TRUE}.}

\item{\code{SNFDistance}}{Which SNF distance calculated cluster assignment
should be included in the plot comparing different methods. To compare
the cluster assignments between SNF distances, an SNF-only Sankey plot
can be created, see parameter \code{AllPlots}. Must be one of:
\code{"euclidean squared"} (default), \code{"euclidean"}, \code{"manhattan"},
\code{"minkowksi 0.25"}, \code{"minkowski 0.5"}, \code{"minkowski 3"} or \code{"minkowski 4"}.}

\item{\code{AllPlots}}{Whether 3 plots should be created per cluster number:
\itemize{
\item Selected: All methods + specified \code{SNFDistance} cluster assignments.
\item SNF: All SNF distance cluster assignments.
\item Full: All methods cluster assignments with all SNF distances.
Default is \code{TRUE}. If \code{FALSE}, only the first option, with all methods
and the specified \code{SNFDistance} will be created.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Top_Feature_Weights"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Top_Feature_Weights}{}}}
\subsection{Method \code{plot_Top_Feature_Weights()}}{
Plot the features with the highest absolute weights in each factor
(component), per omics type and multi-omics integration method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Top_Feature_Weights(
  nFeatures = 10,
  Scale = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nFeatures}}{Number of features to plot for each factor. Default is
10.}

\item{\code{Scale}}{Whether to scale the data from the different omics types, to
make the weights more comparable between omics types. Default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Dendrogram"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Dendrogram}{}}}
\subsection{Method \code{plot_Dendrogram()}}{
Plot all previously created dendrograms with metadata features
labeled in colored bars.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Dendrogram(
  Data,
  Clusters = NULL,
  MetadataFeatures,
  LabelMetadata,
  MetadataHeight = 0,
  LabelSize = 1,
  BarLabelSize = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Data}}{Which previously created dendrogram to plot. Must be one of:
\itemize{
\item \code{"Single"}: Single omics sample hierarchical clustering results. Will
plot the dendrograms for all omics types provided.
\item \code{"Ensemble MDS"}: Ensemble MDS sample hierarchical clustering results.
\item \code{"Ensemble Factor"}: Ensemble factor hierarchical clustering results,
as calculated with \code{run_Feature_Ensemble_HClust} with \code{Method}
\code{"Concatenation"}.
\item \code{"Ensemble Feature Average"}: Ensemble feature hierarchical
clustering results, as calculated with \code{run_Feature_Ensemble_HClust}
with \code{Method} \code{"Average"}.
\item \code{"Ensemble Feature Concatenation"}: Ensemble feature hierarchical
clustering results, as calculated with \code{run_Feature_Ensemble_HClust}
with \code{Method} \code{"Concatenation"}.
\item \code{"Multi-Omics Feature"}: Feature hierarchical clustering results per
multi-omics integration method, as calculated with
\code{run_Multi_Omics_Feature_HClust}.
}}

\item{\code{Clusters}}{Number of clusters to color the dendrogram branches for.
If \code{NULL}, the branches will be black.}

\item{\code{MetadataFeatures}}{Names of the sample metadata features to be
included in colored bars below the dendrogram. A string vector with the
names of the metadata features as in the metadata files should be
provided. Can be used in \code{Data} = \code{"Single"} and \code{Data} =
\code{"Ensemble MDS"}.}

\item{\code{LabelMetadata}}{Name of the metadata feature for which the sample
labels should be colored. Must be a sting with the name of the metadata
feature as in the metadata files. Can be used in \code{Data} = \code{"Single"} and
\code{Data} = \code{"Ensemble MDS"}.}

\item{\code{MetadataHeight}}{Height of the colored bars with metadata
information. Default is 0. If 0 puts the colored bars in the dendrogram,
choose a negative integer. For \code{Data} = \code{"Ensemble MDS"},
\code{"Ensemble Feature Average"} or \code{"Ensemble Feature Concatenation"}, a
single integer should be provided. For \code{Data} = \code{"Single"} or
\code{"Multi-Omics Feature"}, a vector of integers with the same length as the
number of omics datasets must be provided.}

\item{\code{LabelSize}}{Size of the label beneath the dendrogram. Default is 1.}

\item{\code{BarLabelSize}}{Size of the label of the colored bars. Default is 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-plot_Enrich_Dot"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-plot_Enrich_Dot}{}}}
\subsection{Method \code{plot_Enrich_Dot()}}{
Plots a dotplot with the results from the overrepresentation analysis
performed in \code{run_Over_Represenation}. Creates a dotplot with the
overrepresented terms per omics type for each cluster of the ensemble
feature clustering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$plot_Enrich_Dot(
  Method = "Average",
  Categories = 10,
  FontSize = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Method}}{Which ensemble feature clustering result is to be used.
Must be one of" \code{"Average"} or \code{"Concatenation"}. Default is \code{"Average"}.}

\item{\code{Categories}}{Maximum number of overrepresented GO terms to be
displayed per cluster. Default is 10.}

\item{\code{FontSize}}{Size of the font used in the plot. Default is 10.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ensemble_Integration_Suite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ensemble_Integration_Suite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensemble_Integration_Suite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
