% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Feature_Concat_HClust}
\alias{Feature_Concat_HClust}
\title{Ensemble feature and factor clustering using concatenated feature weight
matrix}
\usage{
Feature_Concat_HClust(
  FullFeatureData,
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2"
)
}
\arguments{
\item{FullFeatureData}{Concatenated matrix of scaled feature weights from all
multi-omics integration methods as calculated by \code{Concat_Feature_Matrix()}.}

\item{Distance}{Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{MinkowskiPower}{Power of the Minkowski distance. Default is \code{NULL}.}

\item{Linkage}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"} (default), \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\value{
A list of results. Concatenated feature distance matrix, stored in
$FeatureDist. Concatenated factor distance matrix, stored in $FactorDist.
Feature hierarchical clustering tree, stored in $FeatureHClust. Factor
hierarchical clustering tree, stored in $FactorHClust. Feature dendrogram,
stored in $FeatureDendro. Factor dendrogram, stored in $FactorDendro.
}
\description{
Create an ensemble feature clustering result by combining the feature weights
from the multi-omics integration methods. Single matrix of scaled feature
weights from all multi-omics integration methods as calculated by
\code{Concat_Feature_Matrix()} is needed. Feature and factor distances are
calculated on this matrix. Hierarchical clustering is performed for the
feature and factor distances.
}
\keyword{internal}
