% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Feature_Dendro_Clustering}
\alias{Feature_Dendro_Clustering}
\title{Cut feature or factor dendrograms into clusters}
\usage{
Feature_Dendro_Clustering(FeatHClust, Clusters)
}
\arguments{
\item{FeatHClust}{Hierarchical clustering results for features or factors
as calculated in either \code{Feature_Method_Av_HClust()} or
\code{Feature_Concat_HClust()}.}

\item{Clusters}{Number of clusters in which to cut the feature or factor
hierarchical clustering tree.}
}
\value{
A list of results. Feature or factor cluster assignment, stored in
$Clusters. Feature or factor dendrogram with branches colored for the number
of clusters, stored in $Dendro.
}
\description{
Cut feature of factor dendrograms as created in either
\code{Feature_Method_Av_HClust()} or \code{Feature_Concat_HClust()} to create cluster
assignments.
}
\keyword{internal}
