% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Feature_Method_Av_HClust}
\alias{Feature_Method_Av_HClust}
\title{Ensemble feature clustering using average feature distance between
integration methods}
\usage{
Feature_Method_Av_HClust(
  ScaledFeaturesData,
  nFactors = NULL,
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2"
)
}
\arguments{
\item{ScaledFeaturesData}{List of scaled feature weight dataframes for the
different multi-omics integration methods, by number of factors/components
and omics type, as calculated with \code{Feature_Weight_Scaling()}.}

\item{nFactors}{Number of factors (components) for which the hierarchical
clustering is to be performed. All multi-omics integration methods will need
to have been performed with this number of factors/components.}

\item{Distance}{Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{MinkowskiPower}{Power of the Minkowski distance. Default is \code{NULL}.}

\item{Linkage}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"} (default), \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\value{
A list of results. Average feature distance matrix, stored in
$DistMat. Hierarchical clustering tree, stored in $HClust. Feature dendrogram,
stored in $Dendrogram.
}
\description{
Create an ensemble feature clustering result by combining the feature weights
from the multi-omics integration methods. Scaled feature weight matrices of
different omics types as calculated by \code{Feature_Weight_Scaling} are combined
per multi-omics integration method. Feature distances are calculated per
method, and the average feature distances are calculated between the methods.
Hierarchical clustering is performed on the average feature distance matrix.
}
\keyword{internal}
