% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Feature_Optimal_Clusters}
\alias{Feature_Optimal_Clusters}
\title{Calculate the optimal cluster number for features or factors using the
hierarchical clustering result.}
\usage{
Feature_Optimal_Clusters(HClust, Dist, WeightedClusterStat, MaxClusters)
}
\arguments{
\item{HClust}{Feature or factor hierarchical clustering tree as calculated in
either \code{Feature_Method_Av_HClust()} or \code{Feature_Concat_HClust()}.}

\item{Dist}{Feature or factor distance matrix as calculated in
\code{Feature_Method_Av_HClust()} or \code{Feature_Concat_HClust()}.}

\item{WeightedClusterStat}{The statistics included to calculate the
optimal cluster number. Must be one of: \code{"all"} (default), \code{"noCH"} (all
statistics except \code{"CH"} and \code{"CHsq"}), \code{"PBC"}, \code{"HG"}, \code{"HGSD"},
\code{"ASW"}, \code{"ASWw"}, \code{"CH"}, \code{"R2"}, \code{"CHsq"}, \code{"R2sq"} or \code{"HC"}.}

\item{MaxClusters}{The maximum number of feature or factor clusters to be
considered by the statistics.}
}
\value{
The optimal number of feature or factor clusters.
}
\description{
Calculate the optimal number of features or factors using the hierarchical
clustering trees and distance matrices calculated in either
\code{Feature_Method_Av_HClust()} or \code{Feature_Concat_HClust()}. Uses the
as.clustrange function from the WeightedCluster package.
}
\references{
Studer M (2013). “WeightedCluster Library Manual: A practical
guide to creating typologies of trajectories in the social sciences with R.”
LIVES Working Papers 24.
}
\keyword{internal}
