% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Feature_Proportion}
\alias{Feature_Proportion}
\title{Calculate the feature proportion of the total feature amount per sample}
\usage{
Feature_Proportion(OmicSampleData)
}
\arguments{
\item{OmicSampleData}{Single omics dataset, with features in rows.}
}
\value{
Single omics matrix with proportion of each feature of feature total
in the sample calculated.
}
\description{
Calculate the feature proportion of the total feature amount per sample,
used in EINS R6 class function \code{run_Preprocessing}, if \code{FunctionOrder}
includes \code{"FeatureProp"}.
}
\keyword{internal}
