% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Feature_Weight_Scaling}
\alias{Feature_Weight_Scaling}
\title{Scale feature weight data from multi-omics integration}
\usage{
Feature_Weight_Scaling(FeatureData)
}
\arguments{
\item{FeatureData}{List of feature weight dataframes for the different
multi-omics integration methods, by number of factors/components and omics
type.}
}
\value{
List of scaled feature weight dataframes, stored by multi-omics
integration method, number of factors/components and omics type.
}
\description{
Scale the feature weight data calculated by the multi-omics integration
methods, per method, number of factors/components and omics type. Scale used
is 0-1. Used in EINS R6 class functions \code{run_Multi_Omics_Feature_HClust} and
\code{run_Feature_Ensemble_HClust}.
}
\keyword{internal}
