% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{filter_Blank_Mean}
\alias{filter_Blank_Mean}
\title{Filter the features based on the measurement in blank samples}
\usage{
filter_Blank_Mean(
  OmicSampleData,
  OmicBlankData,
  BlankMeanMultiplier = 2,
  SampleThreshold = 0.8
)
}
\arguments{
\item{OmicSampleData}{Single omics dataset, with features in rows.}

\item{OmicBlankData}{Single omics blank samples dataset, with features in
rows.}

\item{BlankMeanMultiplier}{Numerical by which the blank feature mean is
multiplied. Default is 2.}

\item{SampleThreshold}{Numerical threshold for blank mean filtering. Default
is 0.8 (at least 0.8 of the samples are larger than the blank feature mean *
\code{BlankMeanMultiplier}).}
}
\value{
Single omics matrix with features with more samples below the
multiplied blank mean than the threshold allows for filtered out.
}
\description{
Filter the features based on the measurement in blank samples. Multiply the
mean feature measurement from blank samples by a user-defined multiplier, and
filter out the features in which fewer features than the user-defined
threshold exceed the multiplied blank mean. Only possible if blank samples
are available. Used in EINS R6 class function \code{run_Preprocessing}, if
\code{FunctionOrder} includes \code{"FilterBlankMean"}.
}
\keyword{internal}
