% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{filter_Coverage}
\alias{filter_Coverage}
\title{Filter features with NA proportions above user-defined threshold}
\usage{
filter_Coverage(OmicSampleData, Coverage = 0.8)
}
\arguments{
\item{OmicSampleData}{Single omics dataset, with features in rows.}

\item{Coverage}{Numerical threshold for feature coverage. Default is 0.8
(at least 0.8 of the samples are not NA per feature).}
}
\value{
Single omics matrix with features with more NAs than the
threshold allows for filtered out.
}
\description{
Filter features based on their proportion of NAs per feature using a
user-defined threshold. Used in EINS R6 class function \code{run_Preprocessing},
if \code{FunctionOrder} includes \code{"FilterCoverage"}.
}
\keyword{internal}
