% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{find_delim}
\alias{find_delim}
\title{Determine file delimiter}
\usage{
find_delim(path)
}
\arguments{
\item{path}{File path name for the data file.}
}
\value{
Vector with file delimiter.
}
\description{
Determine file delimiter of file in file path, used in EINS R6 class function
\code{add_Omics_file}.
}
\references{
Olivier-Jimenez, D., Derks, R.J.E., Harari, O., Cruchaga, C.,
Ali, M., Ori, A., Di Fraia, D., Cabukusta, B., Henrie, A., Giera, M., and
Mohammed, Y. (2025). iSODA: A Comprehensive Tool for Integrative Omics Data
Analysis in Single- and Multi-Omics Experiments. Analytical Chemistry,
Vol. 97, Issue 5, Pages 2689-2697.
}
\keyword{internal}
