% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{getLRAcluster_MOVICS}
\alias{getLRAcluster_MOVICS}
\title{Wrapper function for LRAcluster, as used in getLRAcluster from the MOVICS
package.}
\usage{
getLRAcluster_MOVICS(
  data = NULL,
  N.clust = NULL,
  type = rep("gaussian", length(data)),
  clusterAlg = "ward.D"
)
}
\arguments{
\item{data}{List of omics matrices, with samples in columns.}

\item{N.clust}{Number of sample clusters to be created.}

\item{type}{String indicating the data type in the list of matrices. Must be
one of: \code{"gaussian"} (default), \code{"binomial"}, or \code{"possion"}.}

\item{clusterAlg}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"} (default), \code{"ward.D2"}, \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\value{
A list of results. $fit, the object return by \code{LRAcluster_MOVICS}.
$clust.res, a dataframe with the sample cluster results. $clust.dend, a
dendrogram of the sample clustering. $mo.method, a string value indicating
the method used for multi-omics integrative clustering
}
\description{
Wrapper function for LRAcluster, as used in getLRAcluster from the MOVICS
package. Used in the EINS R6 class function \code{run_LRAcluster}.
}
\references{
Lu, X., Meng, J., Zhou, Y., Jiang, L., and Yan, F. (2020).
MOVICS: an R package for multi-omics integration and visualization in cancer
subtyping. Bioinformatics, btaa1018.
}
\keyword{internal}
