% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Heatmap_Cluster}
\alias{Heatmap_Cluster}
\title{Interactive heatmap with sample clustering result}
\usage{
Heatmap_Cluster(
  Data,
  SampleClusters = NULL,
  FeatureClusters = TRUE,
  YFontSize = 5,
  XFontSize = 5,
  Title = NULL
)
}
\arguments{
\item{Data}{Omics matrix with features in rows and samples in columns.}

\item{SampleClusters}{Dataframe with sample clustering result with samples in
rows and cluster assignment in the column.}

\item{FeatureClusters}{Whether to order the features on the Y-axis based on
hierarchical clustering. Default is \code{TRUE}. If \code{FALSE}, the features will be
ordered as in \code{Data}.}

\item{YFontSize}{Font size of the feature labels on the Y-axis. Default is 5.}

\item{XFontSize}{Font size of the sample labels on the X-axis. Default is 5.}

\item{Title}{Title for the heatmap.}
}
\value{
Interactive heatmap with cluster assignment on X-axis.
}
\description{
Interactive heatmap, with sample clustering result.
}
\keyword{internal}
