% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{impute_NA}
\alias{impute_NA}
\title{Missing value (NA) imputation}
\usage{
impute_NA(OmicSampleData, Method = NULL, K_Neighbors = 10)
}
\arguments{
\item{OmicSampleData}{Single omics dataset, with features in rows.}

\item{Method}{String with method of NA imputation. Must be one of: \code{"KNN"},
\code{"missForest"}, \code{"zero replacement"}, or \code{"omit NA"}.}

\item{K_Neighbors}{Numerical, number of k nearest neighbors. Only required
when \code{Method} = \code{"KNN"}.}
}
\value{
Single omics matrix with NA values imputed, replaced, or features
omitted.
}
\description{
NA imputation or replacement with one of the following methods: k-nearest
neighbor imputation, missForest imputation, zero replacement or NA omittance.
Used in EINS R6 class function \code{run_Preprocessing}, if \code{FunctionOrder}
includes \code{"NAImpute"}.
}
\references{
\itemize{
\item Hastie T, Tibshirani R, Narasimhan B, Chu G (2025). impute:
impute: Imputation for microarray data. doi:10.18129/B9.bioc.impute,
R package version 1.82.0
\item Stekhoven DJ (2022). missForest: Nonparametric Missing Value Imputation
using Random Forest. R package version 1.5.
}
}
\keyword{internal}
