% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{LRAcluster_MOVICS}
\alias{LRAcluster_MOVICS}
\title{Integrated analysis of cancer omics data by low-rank approximation using
LRAcluster code from the MOVICS package}
\usage{
LRAcluster_MOVICS(
  data,
  types,
  dimension = 2,
  names = as.character(1:length(data))
)
}
\arguments{
\item{data}{List of omics matrices, with samples in columns.}

\item{types}{List of data types, must be one of \code{"gaussian"} (default),
\code{"binomial"}, or \code{"poisson"}.}

\item{dimension}{The reduced dimension. Default is 2.}

\item{names}{Names of the datasets.}
}
\value{
A list of results. $coordinate, a matrix of the coordinates of the
samples in the reduced space. $potential, the ratio of explained variance.
}
\description{
LRAcluster function from the MOVICS package. Used in the wrapper function
\code{getLRAcluster_MOVICS}.
}
\references{
\itemize{
\item Lu, X., Meng, J., Zhou, Y., Jiang, L., and Yan, F. (2020).
MOVICS: an R package for multi-omics integration and visualization in cancer
subtyping. Bioinformatics, btaa1018.
\item Wu D, Wang D, Zhang MQ, Gu J (2015). Fast dimension reduction and
integrative clustering of multi-omics data using low-rank approximation:
application to cancer molecular classification. BMC Genomics, 16(1):1022.
}
}
\keyword{internal}
