% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Match_Metadata}
\alias{Match_Metadata}
\title{Combine sample metadata from different omics for samples matches across omics}
\usage{
Match_Metadata(MetadataList)
}
\arguments{
\item{MetadataList}{List of metadata files sorted by omics.}
}
\value{
Dataframes with matching metadata in matching samples.
}
\description{
Create a dataframe with identical metadata between different omics metadata
files, for the samples which exist across all omics.
}
\keyword{internal}
