% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Multi_Omics_Feature_Clustering}
\alias{Multi_Omics_Feature_Clustering}
\title{Cluster scaled feature weight data from multi-omics integration}
\usage{
Multi_Omics_Feature_Clustering(
  ScaledFeaturesData,
  nFactors = NULL,
  Distance = "euclidean",
  MinkowskiPower = NULL,
  Linkage = "ward.D2"
)
}
\arguments{
\item{ScaledFeaturesData}{List of scaled feature weight dataframes for the
different multi-omics integration methods, by number of factors/components
and omics type, as calculated with \code{Feature_Weight_Scaling()}.}

\item{nFactors}{Number of factors (components) for which the hierarchical
clustering is to be performed. All multi-omics integration methods will need
to have been performed with this number of factors/components.}

\item{Distance}{Distance metric to be used for the calculation of the
feature distance matrices. Must be one of: \code{"euclidean"} (default),
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
If \code{"minkowski"} is selected, argument \code{"MinkowskiPower"} needs to be
included.}

\item{MinkowskiPower}{Power of the Minkowski distance. Default is \code{NULL}.}

\item{Linkage}{Agglomeration method to be used for the hierarchical
clustering. Must be one of: \code{"ward.D"}, \code{"ward.D2"} (default), \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\value{
A list of results. Feature distance matrices per multi-omics
integration method, stored in $DistMat. Hierarchical clustering trees per
multi-omics integration method, stored in $HClust. Feature dendrogram per
multi-omics integration method, stored in $Dendrogram.
}
\description{
Use scaled feature weight data per multi-omics integration method as
calculated by \code{Feature_Weight_Scaling()} to calculate a feature distance
matrix per multi-omics integration method. Hierarchical clustering is then
performed on these matrices to create a dendrogram of features per
multi-omics integration method.
}
\keyword{internal}
