% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{MultiOmicsHeatmapClustered}
\alias{MultiOmicsHeatmapClustered}
\title{Non-interactive heatmap with cluster annotation for all omics in single plot}
\usage{
MultiOmicsHeatmapClustered(
  OmicsData,
  ClusterRes,
  LegendNames = NULL,
  MetadataColumn = NULL,
  MetaData = NULL
)
}
\arguments{
\item{OmicsData}{List of omics matrices, with features in rows and samples
in columns.}

\item{ClusterRes}{Dataframe with a sample cluster assignment.}

\item{LegendNames}{Names for the legends of the different omics heatmaps.
If \code{NULL}, the omics names as used in the EINS object are used. Otherwise,
a string vector with the same length as \code{OmicsData} should be provided.}

\item{MetadataColumn}{Names of the sample metadata features to be displayed
on the heatmaps. These metadata features need to be available in the metadata
files for all omics types. A string vector with the names of the metadata
features as in \code{MetaData} should be provided.}

\item{MetaData}{List of metadata dataframes for all omics.}
}
\value{
Multi-omics heatmap with cluster assignment and metadata features
on the X-axis.
}
\description{
A non-interactive heatmap with all omics in a single plot, and multi-omics
integration or ensemble sample cluster assignment as well as sample metadata
features plotted on the X-axis.
}
\keyword{internal}
