% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{normalize_ZScore}
\alias{normalize_ZScore}
\title{Feature Z-score normalization}
\usage{
normalize_ZScore(OmicSampleData)
}
\arguments{
\item{OmicSampleData}{Single omics dataset, with features in rows.}
}
\value{
Single omics matrix with features normalized with Z-score
normalization.
}
\description{
Perform Z-score normalization on the features. Used in EINS R6 class function
\code{run-Preprocessing}, if \code{FunctionOrder} includes \code{"Normalization"}.
}
\keyword{internal}
