% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Optimal_Cluster_Number}
\alias{Optimal_Cluster_Number}
\title{Calculate optimal cluster number of integrated multi-omics data with NbClust.}
\usage{
Optimal_Cluster_Number(
  Data = NULL,
  Components = NULL,
  NbClustDistance = "euclidean",
  NbClustMinClust = 2,
  NbClustMaxClust = 12,
  NbClustMethod = "kmeans",
  NbClustIndex = "all",
  NbClustAlphaBeale = 0.1
)
}
\arguments{
\item{Data}{List of omics matrices, with samples in columns.}

\item{Components}{Numerical or range of the components/factors to perform
the integrations with.}

\item{NbClustDistance}{String with the distance measure to be used for
the NbClust dissimilarity matrix computation. Must be one of:
\code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or
\code{"minkowski"}. Default is \code{"euclidean"}.}

\item{NbClustMinClust}{Minimal number of clusters considered by NbClust.
Value between 1 and (number of objects - 1). Default is 2.}

\item{NbClustMaxClust}{Maximum number of clusters considered by NbClust.
Value between 2 and (number of object -2), but larger than
NbClustMinClust. Default is 12.}

\item{NbClustMethod}{The cluster analysis method used by NbClust. Must be
one of: \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"},
\code{"mcquitty"}, \code{"median"}, \code{"centroid"}, \code{"kmeans"}. Default is
\code{"ward.D2"}.}

\item{NbClustIndex}{Indexes to be calculated by NbClust. Must be one of:
\code{"kl"}, \code{"ch"}, \code{"hartigan"}, \code{"ccc"}, \code{"scott"}, \code{"marriot"}, \code{"trcovw"},
\code{"tracew"}, \code{"friedman"}, \code{"rubin"}, \code{"cindex"}, \code{"db"}, \code{"silhouette"},
\code{"duda"}, \code{"pseudot2"}, \code{"beale"}, \code{"ratkowsky"}, \code{"ball"},
\code{"ptbiserial"}, \code{"gap"}, \code{"frey"}, \code{"mcclain"}, \code{"gamma"}, \code{"gplus"},
\code{"tau"}, \code{"dunn"}, \code{"hubert"}, \code{"sdindex"}, \code{"dindex"}, \code{"sdbw"}, \code{"all"}
(all indices except GAP, Gamma, Gplus and Tau), \code{"alllong"} (all indices
with Gap, Gamma, Gplus and Tau included). Default is \code{"all"}.}

\item{NbClustAlphaBeale}{Numerical, significance value for Beale's index.
Default is 0.1.}
}
\value{
A list of the optimal cluster number calculated by each included
index for each integration method per component/factor. The mean, median
and mode are calculated per integration method, and for all methods
combined.
}
\description{
Calculate the optimal cluster number for integrated data using NbClust.
Performs data integration with seven methods that provide sample by
factor/component matrices: MoCluster, MCIA, jNMF, iNMF, LRAcluster, MOFA and
GAUDI. Integration methods are performed with default parameters. A range of
factors/components can be provided, for multiple integration results to be
used. NbClust is then performed on these matrices. Multiple NbClust indexes
are calculated per matrix. All results from NbClust indexes for all matrices
are combined and the mean, mode and median optimal cluster number are
calculated. The user can then determine the preferred cluster number. Used in
EINS R6 class function \code{run_Optimal_Cluster_Number}.
}
\references{
\itemize{
\item Charrad M, Ghazzali N, Boiteau V, Niknafs A (2014). “NbClust:
An R Package for Determining the Relevant Number of Clusters in a Data Set.”
Journal of Statistical Software, 61(6), 1–36.
\item Meng C (2025). mogsa: Multiple omics data integrative clustering and gene
set analysis. doi:10.18129/B9.bioc.mogsa, R package version 1.42.0
\item Meng C, Kuster B, Culhane A, Gholami AM (2013). “A multivariate approach
to the integration of multi-omics datasets.” BMC Bioinformatics.
\item Tsuyuzaki, K., and Nikaido, I. (2024). nnTensor: Non-Negative Tensor
Decomposition. R package version 1.3.0
\item Chalise, P., Raghavan, R., and Fridley, B. (2025). IntNMF: Integrative
Clustering of Multiple Genomic Dataset. R package version 1.3.0
\item Lu, X., Meng, J., Zhou, Y., Jiang, L., and Yan, F. (2020). MOVICS: an R
package for multi-omics integration and visualization in cancer subtyping.
Bioinformatics, btaa1018.
\item Wu D, Wang D, Zhang MQ, Gu J (2015). Fast dimension reduction and
integrative clustering of multi-omics data using low-rank approximation:
application to cancer molecular classification. BMC Genomics, 16(1):1022.
\item Argelaguet R, Velten B, Arnol D, Dietrich S, Zenz T, Marioni JC, Buettner
F, Huber W, Stegle O (2018). “Multi‐Omics Factor Analysis—a framework for
unsupervised integration of multi‐omics data sets.” Molecular Systems
Biology, 14.
\item Castellano-Escuder P, Zachman DK, Han K, Hirschey MD. GAUDI:
interpretable multi-omics integration with UMAP embeddings and
density-based clustering. Nat Commun. 2025 Jul 1;16(1):5771.
}
}
\keyword{internal}
