% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{ORA_GO}
\alias{ORA_GO}
\title{Feature overrepresentation analysis on the clustered ensemble feature results}
\usage{
ORA_GO(
  FullList,
  SelectList,
  GeneNameType,
  Subontologies = "ALL",
  pValue = 0.05,
  pAdjustment = "BH",
  qValue = 0.2,
  MinGenesPerTerm = 10,
  MaxGenesPerTerm = 500
)
}
\arguments{
\item{FullList}{Vector of all features of the analyzed omics type.}

\item{SelectList}{Vector of features of the analyzed omics type in the
cluster.}

\item{GeneNameType}{Keytype of the feature name for the omics dataset. Must
be one of: \code{"ENTREZID"}, \code{"PFAM"}, \code{"IPI"}, \code{"PROSITE"}, \code{"ACCNUM"},
\code{"ALIAS"}, \code{"CHR"}, \code{"CHRLOC"}, \code{"CHRLOCEND"}, \code{"ENZYME"}, \code{"MAP"}, \code{"PATH"},
\code{"PMID"}, \code{"REFSEQ"}, \code{"SYMBOL"}, \code{"UNIGENE"}, \code{"ENSEMBL"}, \code{"ENSEMBLPROT"},
\code{"ENSEMBLTRANS"}, \code{"GENENAME"}, \code{"UNIPROT"}, \code{"GO"}, \code{"EVIDENCE"},
\code{"ONTOLOGY"}, \code{"GOALL"}, \code{"EVIDENCEALL"}, \code{"ONTOLOGYALL"}, \code{"OMIM"} or
\code{"UCSCKG"}.}

\item{Subontologies}{Which ontology to use. Must be one of: \code{"ALL"} (default),
\code{"MF"}, \code{"BP"} or \code{"CC"}.}

\item{pValue}{Cutoff value for the p-value. Default is 0.05.}

\item{pAdjustment}{Method for the p-value adjustment. Must be one of: \code{"BH"}
(default), \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"}, \code{"BY"}, \code{"fdr"}
or \code{"none"}.}

\item{qValue}{Cutoff for the q-value. Default is 0.2.}

\item{MinGenesPerTerm}{Minimal number of genes annotated per ontology term to
be included for testing. Default is 10.}

\item{MaxGenesPerTerm}{Maximum number of genes annotated per ontology term to
be included for testing. Default is 500.}
}
\value{
A list of results stored in class \code{"enrichResult"}. Enrichment
analysis result, stored in $result. p-value cutoff value, stored in
$pvalueCutoff. p-value adjustment method, stored in $pAdjustMethod. q-value
cutoff value, stored in $qvalueCutoff. Organism for which enrichment is
performed, can only be \code{"human"}, stored in $organism. Ontology used, stored
in $ontology. Feature IDs in cluster, stored in $gene. Feature keytype,
stored in $keytype. All feature ID's, stored in $universe. Feature and
category association, stored in $geneInCategory. Mapping of feature to
Symbol, stored in $gen2Symbol. Feature sets, stored in $geneSets. Logical
flag of feature ID in symbol or not, stored in $readable.
}
\description{
Perform overrepresentation analysis from package ClusterProfiler on the
clustered ensembled features to determine if features with similar biological
functions group together in across multi-omics integration methods. Uses
enrichment GO categories from the genome wide annotation database for human
as provided in package org.Hs.eg.db. Performed per cluster and per omics type.
Uses feature cluster assignments created in \code{Feature_Dendro_Clustering()}.
}
\references{
\itemize{
\item Yu G (2024). “Thirteen years of clusterProfiler.” The
Innovation, 5(6), 100722.
\item Xu S, Hu E, Cai Y, Xie Z, Luo X, Zhan L, Tang W, Wang Q, Liu B, Wang R,
Xie W, Wu T, Xie L, Yu G (2024). “Using clusterProfiler to characterize
multiomics data.” Nature Protocols, 19(11).
\item Wu T, Hu E, Xu S, Chen M, Guo P, Dai Z, Feng T, Zhou L, Tang W, Zhan L,
Fu x, Liu S, Bo X, Yu G (2021). “clusterProfiler 4.0: A universal
enrichment tool for interpreting omics data.” The Innovation, 2(3).
\item Yu G, Wang L, Han Y, He Q (2012). “clusterProfiler: an R package for
comparing biological themes among gene clusters.” OMICS: A Journal of
Integrative Biology, 16(5), 284-287.
\item Carlson M (2025). "Genome wide annotation for Human, primarily based on
mapping using Entrez Gene identifiers." Bioconductor.
}
}
\keyword{internal}
