% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{OverlapScaling}
\alias{OverlapScaling}
\title{Min max scaling}
\usage{
OverlapScaling(x)
}
\arguments{
\item{x}{Dataframe with single numerical column.}
}
\value{
Scaled dataframe.
}
\description{
Min max scaling into scale 0-1. Used in function \code{Feature_Weight_Scaling}.
}
\keyword{internal}
