% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{PCA_Plot_Batch}
\alias{PCA_Plot_Batch}
\title{PCA plot of sample batch effect}
\usage{
PCA_Plot_Batch(Data, Metadata, BatchColumnName, XAxis = 1, YAxis = 2, Title)
}
\arguments{
\item{Data}{Single omics dataset, with features in rows.}

\item{Metadata}{Single omics metadata file, with batch information.}

\item{BatchColumnName}{String of the column name which contains batch
information in the metadata file.}

\item{XAxis}{Numerical to indicate which principle component will be used for
the X-axis in the PCA plot. Default is 1.}

\item{YAxis}{Numerical to indicate which principle component will be used for
the Y-acis in the PCA plot. Default is 2.}

\item{Title}{String, title for the batch effect PCA plot.}
}
\value{
Interactive scatterplot with samples colored by batch.
}
\description{
Creates a PCA plot of the samples, colored by batch to examine the batch
effect. Only possible if batch information is available for the samples. Used
in EINS R6 class function \code{run_Preprocessing}, if \code{FunctionOrder} includes
\code{"BatchCorrection"}.
}
\keyword{internal}
