% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Plot_Clusters_Metadata}
\alias{Plot_Clusters_Metadata}
\title{Table with cells colored for sample cluster assignment and metadata features}
\usage{
Plot_Clusters_Metadata(MetadataColumns = NULL, MatchedMetadata, ClusterResults)
}
\arguments{
\item{MetadataColumns}{Names of the sample metadata features to be included
in the table. These metadata features need to be available in the metadata
files for all omics types. A string vector with the names of the metadata
features as in the metadata files should be provided.}

\item{MatchedMetadata}{A dataframe of matched metadata features across omics,
as provided by \code{Match_Metadata()}.}

\item{ClusterResults}{Dataframe with sample cluster assignment.}
}
\value{
A cluster assignment and metadata feature table.
}
\description{
Table with columns of sample cluster assignment and metadata features, with
each column having a different color palette.
}
\keyword{internal}
