% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{Sankey_Plot}
\alias{Sankey_Plot}
\title{Sankey plot for sample movement throughout clusters}
\usage{
Sankey_Plot(SankeyList, Title = "Sankey Plot")
}
\arguments{
\item{SankeyList}{List of Sankey plot input, as created in either
\code{Data_Manipulation_Sankey_Clusters()} or \code{Data_Manipulation_Sankey_Methods()}.}

\item{Title}{String indicating the title for the plot. Default is
\code{"Sankey Plot"}.}
}
\value{
Sankey plot displaying samples through different cluster assignments,
with samples colored for a metadata feature, and a legend of this metadata
feature.
}
\description{
Using the list of dataframes created in either
\code{Data_Manipulation_Sankey_Clusters()} or \code{Data_Manipulation_Sankey_Methods()},
create a Sankey plot displaying the sample movement through different cluster
assignments. An interactive Sankey plot with samples colored by a metadata
feature.
}
\keyword{internal}
