% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Scatterplot}
\alias{Scatterplot}
\title{Interactive sample scatterplot}
\usage{
Scatterplot(
  Data,
  XAxis = XAxis,
  YAxis = YAxis,
  Width = 5,
  Height = 5,
  YFontSize = 15,
  XFontSize = 15,
  Title
)
}
\arguments{
\item{Data}{Dataframe with samples in rows, columns with factors
(components/dimensions), a column with cluster assignment and an optional
column with a sample metadata feature for the shape of the dots.}

\item{XAxis}{Integer, factor (component/dimension) to plot on the X-axis.
Default is 1.}

\item{YAxis}{Integer, factor (component/dimension) to plot on the Y-axis.
Default is 2.}

\item{Width}{Width of the scatterplot. Default is 5.}

\item{Height}{Height of the scatterplot. Default is 5.}

\item{YFontSize}{Size of the Y-axis title. Default is 15.}

\item{XFontSize}{Size of the X-axis title. Default is 15.}

\item{Title}{Title of the scatterplot.}
}
\value{
An interactive sample scatterplot, colored by cluster and,
optionally, with shapes determined by a metadata feature.
}
\description{
Interactive scatterplot of samples in the factors (components/dimensions),
colored by cluster.
}
\keyword{internal}
