% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{SNF_Distance_Affinity_Matrix}
\alias{SNF_Distance_Affinity_Matrix}
\title{Calculate affinity matrix as needed for SNF.}
\usage{
SNF_Distance_Affinity_Matrix(
  Distance = NULL,
  Data = NULL,
  Neighbours = NULL,
  Sigma = NULL
)
}
\arguments{
\item{Distance}{String indicating the distance used to calculate the
distance matrix. Must be one of: \code{"euclidean squared"}, \code{"euclidean"},
\code{"manhattan"}, \code{"minkowski 0.25"}, \code{"minkowski 0.5"}, \code{"minkowski 3"},
\code{"minkowski 4"}.}

\item{Data}{List of omics matrices, with samples in rows.}

\item{Neighbours}{Number of nearest neighbors.}

\item{Sigma}{Variance for the local model.}
}
\value{
Affinity matrix calculated with the distance metric provided.
}
\description{
Wrapper function to calculate the affinity matrix needed for SNF. All
distance metrics provided in the EINS method can be used.
}
\references{
Wang, B., Mezlini, A., Demir, F., Fiume, M., Tu, Z., Brudno,
M., Haibe-Kains, B., and Goldenberg, A. (2021) SNFtool: Similarity
Network Fusion. R package version 2.3.1
}
\keyword{internal}
