% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v5.R
\name{soda_read_table}
\alias{soda_read_table}
\title{Import data files from file path}
\usage{
soda_read_table(file_path, sep = NA, first_column_as_index = FALSE)
}
\arguments{
\item{file_path}{File path name for the data file.}

\item{sep}{Separator used in the data file. Default is \code{NA}.}

\item{first_column_as_index}{Whether to use the first column of the data file
as the index. Default is \code{FALSE}.}
}
\value{
Data table from the file.
}
\description{
Import data files from file path, used in EINS R6 class function
\code{add_Omics_file}.
}
\references{
Olivier-Jimenez, D., Derks, R.J.E., Harari, O., Cruchaga, C.,
Ali, M., Ori, A., Di Fraia, D., Cabukusta, B., Henrie, A., Giera, M., and
Mohammed, Y. (2025). iSODA: A Comprehensive Tool for Integrative Omics Data
Analysis in Single- and Multi-Omics Experiments. Analytical Chemistry,
Vol. 97, Issue 5, Pages 2689-2697.
}
\keyword{internal}
