% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EINS_Utils_v7.R
\name{Top_Feature_Weights_All_Factors}
\alias{Top_Feature_Weights_All_Factors}
\title{Plot features with heightest absolute weigths per multi-omics integration
method}
\usage{
Top_Feature_Weights_All_Factors(FeatureRes, NumberFeatures = 10, Scale = TRUE)
}
\arguments{
\item{FeatureRes}{Dataframe with feature weights per factor (component).}

\item{NumberFeatures}{Number of features to plot for each factor. Default is
10.}

\item{Scale}{Whether to scale the data from the different omics types, to
make the weights more comparable between omics types. Default is \code{TRUE}.}
}
\value{
Plot with the top n absolute features by weight, for each factor
(component).
}
\description{
Plot the features with the highest absolute weights in each factor
(component), per omics type and multi-omics integration method.
}
\keyword{internal}
